refactor: split oversized source files into focused units
Break apart the largest production and FVT files along existing
responsibility boundaries so navigation and review stay manageable,
without changing package APIs or runtime behavior.
