+++ internal/eval_hub/storage/sql/evaluations_status.go
@@ -3,10 +3,7 @@ package sql
 import (
 	"database/sql"
 	"encoding/json"
-	"fmt"
-	"reflect"
 
-	"github.com/eval-hub/eval-hub/internal/eval_hub/abstractions"
 	"github.com/eval-hub/eval-hub/internal/eval_hub/constants"
 	"github.com/eval-hub/eval-hub/internal/eval_hub/handlers"
 	"github.com/eval-hub/eval-hub/internal/eval_hub/messages"
@@ -14,119 +11,6 @@ import (
 	"github.com/eval-hub/eval-hub/internal/eval_hub/storage/sql/shared"
 	"github.com/eval-hub/eval-hub/pkg/api"
 )
-
-type EvaluationJobEntity struct {
-	Config  *api.EvaluationJobConfig  `json:"config" validate:"required"`
-	Status  *api.EvaluationJobStatus  `json:"status,omitempty"`
-	Results *api.EvaluationJobResults `json:"results,omitempty"`
-}
-
-// #######################################################################
-// Evaluation job operations
-// #######################################################################
-func (s *sqlStorage) CreateEvaluationJob(evaluation *api.EvaluationJobResource) error {
-	return s.withTransaction("create evaluation job", evaluation.Resource.ID, func(txn *sql.Tx) error {
-		evaluationJSON, err := s.createEvaluationJobEntity(evaluation)
-		if err != nil {
-			return se.WithRollback(err)
-		}
-		addEntityStatement, args := s.statementsFactory.CreateEvaluationAddEntityStatement(evaluation, string(evaluationJSON))
-		_, err = s.exec(txn, addEntityStatement, args...)
-		if err != nil {
-			return se.WithRollback(err)
-		}
-		s.logger.Info("Created evaluation job", "id", evaluation.Resource.ID, "addEntityStatement", addEntityStatement)
-		return nil
-	})
-}
-
-func (s *sqlStorage) createEvaluationJobEntity(evaluation *api.EvaluationJobResource) ([]byte, error) {
-	evaluationEntity := &EvaluationJobEntity{
-		Config:  &evaluation.EvaluationJobConfig,
-		Status:  evaluation.Status,
-		Results: evaluation.Results,
-	}
-	evaluationJSON, err := json.Marshal(evaluationEntity)
-	if err != nil {
-		return nil, se.NewServiceError(messages.InternalServerError, "Error", err.Error())
-	}
-	return evaluationJSON, nil
-}
-
-func (s *sqlStorage) GetEvaluationJob(id string) (*api.EvaluationJobResource, error) {
-	return s.getEvaluationJobTransactional(nil, id)
-}
-
-func (s *sqlStorage) getEvaluationJobTransactional(txn *sql.Tx, id string) (*api.EvaluationJobResource, error) {
-	return s.scanEvaluationJobTransactional(txn, id, false)
-}
-
-func (s *sqlStorage) getEvaluationJobTransactionalForUpdate(txn *sql.Tx, id string) (*api.EvaluationJobResource, error) {
-	return s.scanEvaluationJobTransactional(txn, id, true)
-}
-
-func (s *sqlStorage) scanEvaluationJobTransactional(txn *sql.Tx, id string, forUpdate bool) (*api.EvaluationJobResource, error) {
-	query := shared.EntityQuery{Resource: api.Resource{ID: id, Tenant: s.tenant}}
-	var selectQuery string
-	var selectArgs, queryArgs []any
-	if forUpdate {
-		selectQuery, selectArgs, queryArgs = s.statementsFactory.CreateEvaluationGetEntityForUpdateStatement(&query)
-	} else {
-		selectQuery, selectArgs, queryArgs = s.statementsFactory.CreateEvaluationGetEntityStatement(&query)
-	}
-
-	err := s.queryRow(txn, selectQuery, selectArgs...).Scan(queryArgs...)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, se.NewServiceError(messages.ResourceNotFound, "Type", "evaluation job", "ResourceId", id)
-		}
-		s.logger.Error("Failed to get evaluation job", "error", err, "id", id)
-		return nil, se.WithRollback(se.NewServiceError(messages.DatabaseOperationFailed, "Type", "evaluation job", "ResourceId", id, "Error", err.Error()))
-	}
-
-	var evaluationJobEntity EvaluationJobEntity
-	err = json.Unmarshal([]byte(query.EntityJSON), &evaluationJobEntity)
-	if err != nil {
-		s.logger.Error("Failed to unmarshal evaluation job entity", "error", err, "id", id)
-		return nil, se.WithRollback(se.NewServiceError(messages.JSONUnmarshalFailed, "Type", "evaluation job", "Error", err.Error()))
-	}
-
-	status := ""
-	job, err := constructEvaluationResource(s.logger, &query, status, &evaluationJobEntity)
-	if err != nil {
-		return nil, se.WithRollback(err)
-	}
-	return job, nil
-}
-
-func (s *sqlStorage) GetEvaluationJobs(filter *abstractions.QueryFilter) (*abstractions.QueryResults[api.EvaluationJobResource], error) {
-	var txn *sql.Tx
-	return listEntities[api.EvaluationJobResource](s, txn, shared.TABLE_EVALUATIONS, filter)
-}
-
-func (s *sqlStorage) DeleteEvaluationJob(id string) error {
-	// Build the DELETE query
-	deleteQuery, args := s.statementsFactory.CreateDeleteEntityStatement(s.tenant, shared.TABLE_EVALUATIONS, id)
-
-	// Execute the DELETE query
-	result, err := s.exec(nil, deleteQuery, args...)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			s.logger.Debug("Evaluation job not found", "id", id)
-			return se.NewServiceError(messages.ResourceNotFound, "Type", "evaluation job", "ResourceId", id)
-		}
-		s.logger.Error("Failed to delete evaluation job", "error", err, "id", id)
-		return se.WithRollback(se.NewServiceError(messages.DatabaseOperationFailed, "Type", "evaluation job", "ResourceId", id, "Error", err.Error()))
-	}
-	if rows, err := result.RowsAffected(); ((err == nil) && (rows == 0)) || (err == sql.ErrNoRows) {
-		s.logger.Debug("Evaluation job not found", "id", id)
-		return se.NewServiceError(messages.ResourceNotFound, "Type", "evaluation job", "ResourceId", id)
-	}
-
-	s.logger.Info("Deleted evaluation job", "id", id)
-
-	return nil
-}
 
 func (s *sqlStorage) checkEvaluationJobState(evaluationJobID string, evaluationJobState api.OverallState, state api.OverallState) (bool, error) {
 	// check if the state is unchanged
@@ -357,31 +241,6 @@ func (s *sqlStorage) updateBenchmarkStatus(job *api.EvaluationJobResource, runSt
 	job.Status.Benchmarks = append(job.Status.Benchmarks, *benchmarkStatus)
 }
 
-func (s *sqlStorage) updateBenchmarkResults(job *api.EvaluationJobResource, runStatus *api.StatusEvent, result *api.BenchmarkResult) error {
-	if job.Results == nil {
-		job.Results = &api.EvaluationJobResults{}
-	}
-	if job.Results.Benchmarks == nil {
-		job.Results.Benchmarks = make([]api.BenchmarkResult, 0)
-	}
-
-	for i, benchmark := range job.Results.Benchmarks {
-		if benchmark.ID == runStatus.BenchmarkStatusEvent.ID &&
-			benchmark.ProviderID == runStatus.BenchmarkStatusEvent.ProviderID &&
-			benchmark.BenchmarkIndex == runStatus.BenchmarkStatusEvent.BenchmarkIndex {
-			if reflect.DeepEqual(benchmark, *result) {
-				return nil
-			}
-			job.Results.Benchmarks[i] = *result
-			return nil
-		}
-	}
-
-	job.Results.Benchmarks = append(job.Results.Benchmarks, *result)
-
-	return nil
-}
-
 // UpdateEvaluationJobWithRunStatus runs in a transaction: fetches the job, merges RunStatusInternal into the entity, and persists.
 func (s *sqlStorage) UpdateEvaluationJob(id string, runStatus *api.StatusEvent) error {
 	return s.withTransaction("update evaluation job", id, func(txn *sql.Tx) error {
@@ -474,154 +333,4 @@ func (s *sqlStorage) UpdateEvaluationJob(id string, runStatus *api.StatusEvent) 
 
 		return nil
 	})
-}
-
-func (s *sqlStorage) computeJobTestResult(job *api.EvaluationJobResource, collection *api.CollectionResource) {
-	if job.Results == nil || job.Results.Benchmarks == nil || len(job.Results.Benchmarks) == 0 {
-		return
-	}
-	var sumOfWeightedScores float32 = 0.0
-	var sumOfWeights float32 = 0.0
-	resolvedJobBenchmarks, err := handlers.GetJobBenchmarks(job, collection)
-	if err != nil {
-		s.logger.Error("Failed to get job benchmarks", "error", err, "job_id", job.Resource.ID)
-		return
-	}
-	for _, benchmark := range job.Results.Benchmarks {
-		if benchmark.Test == nil {
-			// if the benchmark test result is not defined, we skip it
-			// This should never happen, since this method is called only when the overall job status is 'completed'
-			s.logger.Info("Benchmark test result is not defined for benchmark", "benchmark_id", benchmark.ID, "benchmark_index", benchmark.BenchmarkIndex)
-			continue
-		}
-		if benchmark.BenchmarkIndex < 0 || benchmark.BenchmarkIndex >= len(resolvedJobBenchmarks) {
-			s.logger.Warn(
-				"benchmark index out of range for resolved benchmarks",
-				"benchmark_id", benchmark.ID,
-				"benchmark_index", benchmark.BenchmarkIndex,
-				"resolved_count", len(resolvedJobBenchmarks),
-			)
-			continue
-		}
-		benchmarkWeight := resolvedJobBenchmarks[benchmark.BenchmarkIndex].Weight
-		if benchmarkWeight == 0 {
-			// if the benchmark weight is not defined, we set it to 1
-			benchmarkWeight = 1
-		}
-		weightedScore := benchmarkWeight * benchmark.Test.PrimaryScore
-		if primaryScore := resolvedJobBenchmarks[benchmark.BenchmarkIndex].PrimaryScore; primaryScore != nil && primaryScore.LowerIsBetter {
-			weightedScore = benchmarkWeight * (1 - benchmark.Test.PrimaryScore)
-		}
-		sumOfWeightedScores += weightedScore
-		sumOfWeights += benchmarkWeight
-		s.logger.Info("Benchmark test result", "benchmark_id", benchmark.ID, "benchmark_index", benchmark.BenchmarkIndex, "primary_score", benchmark.Test.PrimaryScore, "weighted_score", weightedScore, "benchmark_weight", benchmarkWeight, "sum_of_weighted_scores", sumOfWeightedScores, "sum_of_weights", sumOfWeights)
-	}
-	if sumOfWeights == 0 {
-		s.logger.Warn("No benchmark weights accumulated; cannot compute job score")
-		return
-	}
-	weightedAvgJobScore := sumOfWeightedScores / sumOfWeights
-	s.logger.Info("Weighted average job score", "weighted_avg_job_score", weightedAvgJobScore, "sum_of_weighted_scores", sumOfWeightedScores, "sum_of_weights", sumOfWeights)
-
-	threshold := getPassCriteriaThreshold(job, collection)
-	jobTest := &api.EvaluationTest{
-		Score:     weightedAvgJobScore,
-		Threshold: threshold,
-		Pass:      weightedAvgJobScore >= threshold,
-	}
-
-	job.Results.Test = jobTest
-}
-
-func getPassCriteriaThreshold(job *api.EvaluationJobResource, collection *api.CollectionResource) float32 {
-	if job.PassCriteria != nil && job.PassCriteria.Threshold != nil {
-		return *job.PassCriteria.Threshold
-	}
-	if collection != nil && collection.PassCriteria != nil && collection.PassCriteria.Threshold != nil {
-		return *collection.PassCriteria.Threshold
-	}
-	// this is the hard-coded default pass criteria threshold
-	return 0.5
-}
-
-func (s *sqlStorage) computeBenchmarkTestResult(txn *sql.Tx, job *api.EvaluationJobResource, benchmarkStatusEvent *api.BenchmarkStatusEvent, collection *api.CollectionResource) *api.BenchmarkTest {
-	// job could have benchmarks array or it could have collection. If it has collection, we need to get the benchmarks from the collection
-	benchmarks, err := handlers.GetJobBenchmarks(job, collection)
-	if err != nil {
-		s.logger.Error("Failed to get job benchmarks", "error", err, "job_id", job.Resource.ID)
-		return nil
-	}
-	if len(benchmarks) == 0 {
-		return nil
-	}
-	for _, benchmark := range benchmarks {
-		if benchmark.ID != benchmarkStatusEvent.ID || benchmark.ProviderID != benchmarkStatusEvent.ProviderID {
-			continue
-		}
-		primaryScore := benchmark.PrimaryScore
-		var providerBench *api.BenchmarkResource
-		// if the primary score is not defined, we need to get the primary score from the provider
-		if (primaryScore == nil || primaryScore.Metric == "") && benchmark.ProviderID != "" {
-			provider, err := s.getUserProviderTransactional(txn, benchmark.ProviderID)
-			if err == nil && provider != nil {
-				for i := range provider.Benchmarks {
-					if provider.Benchmarks[i].ID == benchmark.ID {
-						providerBench = &provider.Benchmarks[i]
-						break
-					}
-				}
-			}
-			if providerBench != nil && providerBench.PrimaryScore != nil && providerBench.PrimaryScore.Metric != "" {
-				primaryScore = providerBench.PrimaryScore
-			}
-		}
-		if primaryScore != nil && primaryScore.Metric != "" {
-			primaryMetric := primaryScore.Metric
-			if primaryMetricValue, ok := benchmarkStatusEvent.Metrics[primaryMetric]; ok {
-				primaryMetricValueFloat, err := castAnyToFloat32(primaryMetricValue)
-				if err != nil {
-					s.logger.Error("Failed to cast primary metric value to float32", "error", err, "primary_metric", primaryMetric, "primary_metric_value", primaryMetricValue)
-					return nil
-				}
-				var threshold float32
-				if benchmark.PassCriteria != nil && benchmark.PassCriteria.Threshold != nil {
-					threshold = *benchmark.PassCriteria.Threshold
-				} else if providerBench != nil && providerBench.PassCriteria != nil && providerBench.PassCriteria.Threshold != nil {
-					threshold = *providerBench.PassCriteria.Threshold
-				} else {
-					return nil
-				}
-				pass := primaryMetricValueFloat >= threshold
-				if primaryScore.LowerIsBetter {
-					pass = primaryMetricValueFloat <= threshold
-				}
-				return &api.BenchmarkTest{
-					PrimaryScore:       primaryMetricValueFloat,
-					PrimaryScoreMetric: primaryMetric,
-					Threshold:          threshold,
-					Pass:               pass,
-				}
-			}
-		}
-	}
-	return nil
-}
-
-func castAnyToFloat32(primaryMetricValue any) (float32, error) {
-	var primaryMetricValueFloat float32
-	switch v := primaryMetricValue.(type) {
-	case float64:
-		primaryMetricValueFloat = float32(v)
-	case float32:
-		primaryMetricValueFloat = v
-	case int:
-		primaryMetricValueFloat = float32(v)
-	case int32:
-		primaryMetricValueFloat = float32(v)
-	case int64:
-		primaryMetricValueFloat = float32(v)
-	default:
-		return 0, fmt.Errorf("unsupported type: %T for primary metric value", primaryMetricValue)
-	}
-	return primaryMetricValueFloat, nil
 }
