--- tests/features/step_definitions_test.go
@@ -1,2338 +1,0 @@
-package features
-
-import (
-	"bytes"
-	"context"
-	"crypto/tls"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"io"
-	"log"
-	"log/slog"
-	"math"
-	"net"
-	"net/http"
-	"net/url"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"regexp"
-	"runtime/debug"
-	"slices"
-	"strconv"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/Jeffail/gabs/v2"
-	"github.com/PaesslerAG/jsonpath"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/config"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/metrics"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/mlflow"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/runtimes"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/server"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/storage"
-	"github.com/eval-hub/eval-hub/internal/eval_hub/validation"
-	"github.com/eval-hub/eval-hub/internal/logging"
-	"github.com/eval-hub/eval-hub/internal/otel"
-	"github.com/eval-hub/eval-hub/internal/testhelpers"
-	pkgapi "github.com/eval-hub/eval-hub/pkg/api"
-	"github.com/eval-hub/eval-hub/pkg/evalhubclient"
-	"github.com/xeipuuv/gojsonschema"
-
-	"github.com/cucumber/godog"
-	"github.com/modelcontextprotocol/go-sdk/mcp"
-
-	mcpserver "github.com/eval-hub/eval-hub/internal/evalhub_mcp/server"
-)
-
-const (
-	valuePrefix  = "value:"
-	mlflowPrefix = "mlflow:"
-	envPrefix    = "env:"
-	regexpPrefix = "regex:"
-
-	envMetricsURL = "METRICS_URL"
-)
-
-// modelEndpointStatus captures preflight outcome from checkModelEndpoint for steps that gate on connectivity.
-type modelEndpointStatus int
-
-const (
-	modelEndpointUnchecked modelEndpointStatus = iota
-	modelEndpointUnreachable
-	modelEndpointReachable
-)
-
-var (
-	// testConfig to be used throughout all the test suites
-	// for the global configuration
-	apiFeat *apiFeature
-
-	once   sync.Once
-	logger *log.Logger
-
-	modelEndpointConnectivity modelEndpointStatus
-)
-
-type apiFeature struct {
-	baseURL        *url.URL
-	metricsBaseURL *url.URL
-	server         *server.Server
-	httpServer     *http.Server
-	metricsServer  *server.MetricsServer
-	client         *http.Client
-	// MCP-specific fields
-	mcpServer        *mcp.Server
-	mcpClientSession *mcp.ClientSession
-	mcpServerSession *mcp.ServerSession
-}
-
-// this is used for a scenario to ensure that scenarios do not overwrite
-// data from other scenarios...
-type scenarioConfig struct {
-	scenarioName string
-	apiFeature   *apiFeature
-	response     *http.Response
-	body         []byte
-
-	reqHeaders map[string]string
-
-	lastURL    string
-	lastMethod string
-	lastId     string
-
-	// MCP-specific fields
-	mcpToolResult    *mcp.CallToolResult
-	mcpError         error
-	mcpResourceText  string
-	mcpResourceError error
-	mcpPromptResult  *mcp.GetPromptResult
-	mcpPromptError   error
-
-	// assetsSync sync.Mutex
-	assets map[string][]string
-
-	values map[string]string
-
-	// jsonnetHarnessEnv overrides process env in the jsonnet harness only (see jsonnetHarnessJSON).
-	jsonnetHarnessEnv map[string]string
-	// jsonnetHarnessEnvOmit drops keys from the harness env snapshot even when set in the process.
-	jsonnetHarnessEnvOmit []string
-	// jsonnetMlflowEnabled overrides harness.mlflow_enabled when non-nil.
-	jsonnetMlflowEnabled *bool
-	// jsonnetQueueEnabled overrides harness.queue_enabled when non-nil.
-	jsonnetQueueEnabled *bool
-
-	waitDeadline time.Duration
-	waitInterval time.Duration
-
-	// MLflow artifact fetching
-	mlflowArtifactBody  []byte
-	mlflowArtifactError error
-}
-
-func getLogger() *log.Logger {
-	once.Do(func() {
-		if logger == nil {
-			path := filepath.Join("bin", "tests.log")
-			path, err := filepath.Abs(path)
-			if err != nil {
-				panic(logError(fmt.Errorf("Failed to get absolute path: %v", err)))
-			}
-			logOutput, err := os.Create(path)
-			if err != nil {
-				panic(logError(fmt.Errorf("Failed to create log file: %v", err)))
-			}
-			logger = log.New(logOutput, "", log.LstdFlags)
-		}
-	})
-	return logger
-}
-
-func logDebug(format string, a ...any) {
-	fmt.Printf(format, a...)
-	getLogger().Printf(format, a...)
-}
-
-func logError(err error, withStack ...bool) error {
-	if len(withStack) > 0 && withStack[0] {
-		getLogger().Printf("Error: %v\n%s\n", err, string(debug.Stack()))
-	} else {
-		getLogger().Printf("Error: %v\n", err)
-	}
-	return err
-}
-
-func checkBaseURL(uri *url.URL, from string) {
-	if uri == nil {
-		panic("Invalid baseURL: nil from " + from)
-	}
-	if uri.String() == "" {
-		panic("Empty baseURL from  " + from)
-	}
-}
-
-// getMLflowTLSConfig returns TLS config for MLflow clients, respecting MLFLOW_INSECURE env var
-func getMLflowTLSConfig() *tls.Config {
-	insecure := true // default to true for backward compatibility with cluster internal certs
-	if insecureStr := os.Getenv("MLFLOW_INSECURE"); insecureStr != "" {
-		if parsed, err := strconv.ParseBool(insecureStr); err == nil {
-			insecure = parsed
-		}
-	}
-	return &tls.Config{InsecureSkipVerify: insecure} //nolint:gosec
-}
-
-// getMLflowHTTPClient returns an HTTP client configured for MLflow API calls
-func getMLflowHTTPClient() *http.Client {
-	return &http.Client{
-		Timeout: 60 * time.Second,
-		Transport: &http.Transport{
-			TLSClientConfig: getMLflowTLSConfig(),
-		},
-	}
-}
-
-// mlflowBaseURL returns the MLflow base URL from MLFLOW_URL env or default cluster internal URL
-func mlflowBaseURL() string {
-	baseURL := os.Getenv("MLFLOW_URL")
-	if baseURL == "" {
-		baseURL = "https://mlflow.redhat-ods-applications.svc.cluster.local:8443"
-	}
-	return strings.TrimRight(baseURL, "/")
-}
-
-// mlflowWorkspace resolves the MLflow workspace using X_TENANT env → X-Tenant header → "tenant" fallback
-func (tc *scenarioConfig) mlflowWorkspace() string {
-	workspace := os.Getenv("X_TENANT")
-	if workspace == "" {
-		if tenant, ok := tc.reqHeaders["X-Tenant"]; ok && tenant != "" {
-			workspace = tenant
-		} else {
-			workspace = "tenant" // fallback
-		}
-	}
-	return workspace
-}
-
-func isMetricsScrapePath(path string) bool {
-	path = strings.TrimSpace(path)
-	if path == "/metrics" {
-		return true
-	}
-	u, err := url.Parse(path)
-	return err == nil && u.Path == "/metrics"
-}
-
-func joinBaseURL(base *url.URL, path string) string {
-	baseStr := strings.TrimRight(base.String(), "/")
-	if strings.HasPrefix(path, baseStr) {
-		return path
-	}
-	if !strings.HasPrefix(path, "/") {
-		path = "/" + path
-	}
-	return baseStr + path
-}
-
-// resolveMetricsBaseURL returns the base URL for Prometheus scrape requests.
-func resolveMetricsBaseURL(apiBase *url.URL) (*url.URL, error) {
-	// METRICS_URL is set (local or remote/cluster).
-	// Input: METRICS_URL=http://evalhub-metrics.<ns>.svc:8081 (or any valid scrape base).
-	// Behavior: parse and return it; used when the pipeline targets the dedicated metrics port directly.
-	if raw := strings.TrimSpace(os.Getenv(envMetricsURL)); raw != "" {
-		u, err := url.Parse(raw)
-		if err != nil {
-			return nil, fmt.Errorf("invalid METRICS_URL: %w", err)
-		}
-		checkBaseURL(u, raw)
-		return u, nil
-	}
-
-	// Remote/cluster mode without METRICS_URL.
-	// Input: SERVER_URL=https://evalhub.example.com, METRICS_URL unset.
-	// Behavior: return nil so callers can skip @metrics scenarios or error if /metrics is requested
-	// (metrics are not on the kube-rbac-proxy route; the pipeline must set METRICS_URL explicitly).
-	if strings.TrimSpace(os.Getenv("SERVER_URL")) != "" {
-		return nil, nil
-	}
-
-	// Local embedded-server mode (SERVER_URL unset, METRICS_URL unset).
-	// Input: apiBase=http://localhost:8080 (or PORT); local mode serves /metrics on the main router.
-	// Behavior: default scrape base to the API base URL.
-	return apiBase, nil
-}
-
-func scenarioHasTag(sc *godog.Scenario, tag string) bool {
-	for _, t := range sc.Tags {
-		tagName := strings.TrimPrefix(t.Name, "@")
-		if tagName == tag {
-			return true
-		}
-	}
-	return false
-}
-
-func createApiFeature() (*apiFeature, error) {
-	timeout := 60 * time.Second
-	if timeoutStr := os.Getenv("TEST_TIMEOUT"); timeoutStr != "" {
-		if eTimeout, err := strconv.Atoi(timeoutStr); err != nil {
-			logDebug("Invalid TEST_TIMEOUT: %v\n", err.Error())
-		} else {
-			timeout = time.Duration(eTimeout) * time.Second
-		}
-	}
-	client := &http.Client{
-		Timeout: timeout,
-	}
-
-	if serverURL := os.Getenv("SERVER_URL"); serverURL != "" {
-		uri, err := url.Parse(serverURL)
-		if err != nil {
-			return nil, logError(fmt.Errorf("Invalid SERVER_URL: %v", err))
-		}
-		checkBaseURL(uri, serverURL)
-		metricsBase, err := resolveMetricsBaseURL(uri)
-		if err != nil {
-			return nil, logError(err)
-		}
-		apiFeat := &apiFeature{client: client, baseURL: uri, metricsBaseURL: metricsBase}
-
-		// Initialize MCP server even when using remote server
-		logger, _, err := logging.NewLogger()
-		if err != nil {
-			return nil, logError(fmt.Errorf("failed to create logger for MCP: %w", err))
-		}
-		if err := apiFeat.setupMCPServer(logger); err != nil {
-			return nil, logError(fmt.Errorf("failed to setup MCP server for remote testing: %w", err))
-		}
-
-		return apiFeat, nil
-	}
-
-	port := 8080
-	if sport := os.Getenv("PORT"); sport != "" {
-		if eport, err := strconv.Atoi(sport); err != nil {
-			logDebug("Invalid PORT: %v\n", err.Error())
-		} else {
-			port = eport
-		}
-	}
-
-	uri := fmt.Sprintf("http://localhost:%d", port)
-	baseURL, err := url.Parse(uri)
-	if err != nil {
-		panic(logError(fmt.Errorf("Invalid baseURL: %v", err)))
-	}
-	checkBaseURL(baseURL, uri)
-
-	metricsBase, err := resolveMetricsBaseURL(baseURL)
-	if err != nil {
-		return nil, logError(err)
-	}
-
-	apiFeat := &apiFeature{
-		client:         client,
-		baseURL:        baseURL,
-		metricsBaseURL: metricsBase,
-	}
-	if err := apiFeat.startLocalServer(port); err != nil {
-		return nil, err
-	}
-	return apiFeat, nil
-}
-
-// ensureFVTOTELConfig enables OTEL metrics export for embedded FVT servers when Prometheus
-// scraping is configured. HTTP request duration is collected by otelhttp.
-func ensureFVTOTELConfig(serviceConfig *config.Config) {
-	if serviceConfig == nil || !serviceConfig.IsPrometheusEnabled() {
-		return
-	}
-	if serviceConfig.OTEL == nil {
-		serviceConfig.OTEL = &config.OTELConfig{}
-	}
-	serviceConfig.OTEL.Enabled = true
-	serviceConfig.OTEL.EnableMetrics = true
-	serviceConfig.OTEL.ExporterType = otel.ExporterTypeStdout
-}
-
-func (a *apiFeature) startLocalServer(port int) error {
-	logger, _, err := logging.NewLogger()
-	if err != nil {
-		return err
-	}
-	validate, err := validation.NewValidator()
-	if err != nil {
-		return logError(err)
-	}
-	version, err := testhelpers.RepoVersion()
-	if err != nil {
-		return logError(err)
-	}
-	serviceConfig, err := config.LoadConfig(logger, version, "local", time.Now().Format(time.RFC3339), "")
-	if err != nil {
-		return logError(fmt.Errorf("failed to load service config: %w", err))
-	}
-	serviceConfig.Service.Port = port
-	serviceConfig.Service.LocalMode = true // set local mode for testing
-
-	// set up the provider configs
-	providerConfigs, err := config.LoadProviderConfigs(logger, validate)
-	if err != nil {
-		// we do this as no point trying to continue
-		return logError(fmt.Errorf("failed to load provider configs: %w", err))
-	}
-
-	if len(providerConfigs) == 0 {
-		return logError(fmt.Errorf("no provider configs loaded"))
-	}
-
-	logger.Info("Providers loaded.")
-	for key := range providerConfigs {
-		providerCfg := providerConfigs[key]
-		if providerCfg.Runtime == nil {
-			return logError(fmt.Errorf("provider %q has no runtime configuration", providerCfg.Resource.ID))
-		}
-		if providerCfg.Runtime.Local == nil {
-			providerCfg.Runtime.Local = &pkgapi.LocalRuntime{}
-		}
-		providerConfigs[key] = providerCfg
-	}
-
-	// set up the collection configs
-	collectionConfigs, err := config.LoadCollectionConfigs(logger, validate)
-	if err != nil {
-		return logError(fmt.Errorf("failed to load collection configs: %w", err))
-	}
-
-	ensureFVTOTELConfig(serviceConfig)
-	if serviceConfig.IsOTELEnabled() {
-		if _, err := otel.SetupOTEL(context.Background(), serviceConfig.OTEL, logger, serviceConfig.IsPrometheusEnabled()); err != nil {
-			return logError(fmt.Errorf("failed to setup OTEL: %w", err))
-		}
-	}
-	if serviceConfig.IsOTELMetricsEnabled() {
-		if err := metrics.Init(); err != nil {
-			return logError(fmt.Errorf("failed to initialize OTEL metrics: %w", err))
-		}
-	}
-
-	storage, err := storage.NewStorage(
-		serviceConfig.Database,
-		collectionConfigs,
-		providerConfigs,
-		serviceConfig.IsOTELStorageScansEnabled(),
-		serviceConfig.IsOTELMetricsEnabled(),
-		logger,
-	)
-	if err != nil {
-		return logError(fmt.Errorf("failed to create storage: %w", err))
-	}
-	logger.Info("Storage created.")
-
-	runtime, err := runtimes.NewRuntime(logger, serviceConfig)
-	if err != nil {
-		return logError(fmt.Errorf("failed to create runtime: %w", err))
-	}
-
-	mlflowClient, err := mlflow.NewMLFlowClient(serviceConfig, logger)
-	if err != nil {
-		return logError(fmt.Errorf("failed to create MLFlow client: %w", err))
-	}
-
-	a.server, err = server.NewServer(logger,
-		serviceConfig,
-		storage,
-		validate,
-		runtime,
-		mlflowClient)
-	if err != nil {
-		return err
-	}
-
-	// Create a test server
-	handler, err := a.server.SetupRoutes()
-	if err != nil {
-		return err
-	}
-	a.httpServer = &http.Server{
-		Addr:    fmt.Sprintf(":%d", port),
-		Handler: handler,
-	}
-
-	// Start server in background
-	listener, err := net.Listen("tcp", fmt.Sprintf(":%d", port))
-	if err != nil {
-		return err
-	}
-
-	go func() {
-		_ = a.httpServer.Serve(listener)
-	}()
-
-	if serviceConfig.IsPrometheusEnabled() {
-		a.metricsServer = server.NewMetricsServer(logger, serviceConfig.Prometheus)
-		go func() {
-			if err := a.metricsServer.Start(); err != nil {
-				logger.Error("Metrics server failed", "error", err.Error())
-			}
-		}()
-	}
-
-	// Initialize MCP server with real eval-hub client
-	if err := a.setupMCPServer(logger); err != nil {
-		return logError(fmt.Errorf("failed to setup MCP server: %w", err))
-	}
-
-	return nil
-}
-
-func (a *apiFeature) setupMCPServer(logger *slog.Logger) error {
-	// Create real eval-hub client pointing to local test server
-	tenant := os.Getenv("X_TENANT")
-	token := os.Getenv("AUTH_TOKEN")
-
-	logger.Info("Setting up MCP server", "base_url", a.baseURL.String(), "tenant", tenant, "has_token", token != "")
-
-	evalhubClient := evalhubclient.NewClient(a.baseURL.String())
-	if tenant != "" {
-		evalhubClient = evalhubClient.WithTenant(tenant)
-	}
-	if token != "" {
-		evalhubClient = evalhubClient.WithToken(token)
-	}
-
-	// Create MCP server with real backend
-	version, err := testhelpers.RepoVersion()
-	if err != nil {
-		version = "unknown" // Fallback for test environment
-	}
-
-	// Get git hash from repo (matches Makefile: git rev-parse --short HEAD)
-	gitHash := "unknown"
-	cmd := exec.Command("git", "rev-parse", "--short", "HEAD")
-	if output, err := cmd.Output(); err == nil {
-		gitHash = strings.TrimSpace(string(output))
-	}
-
-	serverInfo := &mcpserver.ServerInfo{
-		Build:     version,
-		BuildDate: time.Now().Format(time.RFC3339),
-		GitHash:   gitHash,
-	}
-
-	a.mcpServer = mcpserver.New(serverInfo, logger, nil)
-	if err := mcpserver.RegisterHandlers(a.mcpServer, evalhubClient, serverInfo, logger, evalhubclient.DefaultListPageLimit); err != nil {
-		return fmt.Errorf("failed to register MCP handlers: %w", err)
-	}
-
-	// Create in-memory transports for testing (like unit tests)
-	serverTransport, clientTransport := mcp.NewInMemoryTransports()
-
-	ctx := context.Background()
-
-	// Connect server
-	serverSession, err := a.mcpServer.Connect(ctx, serverTransport, nil)
-	if err != nil {
-		return fmt.Errorf("failed to connect MCP server: %w", err)
-	}
-	a.mcpServerSession = serverSession
-
-	// Connect client
-	mcpClient := mcp.NewClient(&mcp.Implementation{Name: "fvt-test-client", Version: "1.0.0"}, nil)
-	clientSession, err := mcpClient.Connect(ctx, clientTransport, nil)
-	if err != nil {
-		return fmt.Errorf("failed to connect MCP client: %w", err)
-	}
-	a.mcpClientSession = clientSession
-
-	logger.Info("MCP server initialized successfully for FVT tests")
-	return nil
-}
-
-func (a *apiFeature) cleanup(ctx context.Context, _ *godog.Scenario, _ error) (context.Context, error) {
-	if a.mcpClientSession != nil {
-		if err := a.mcpClientSession.Close(); err != nil {
-			// Log but don't fail - consistent with existing cleanup pattern
-			logDebug("MCP client session close error (non-fatal): %v\n", err)
-		}
-	}
-	if a.mcpServerSession != nil {
-		if err := a.mcpServerSession.Close(); err != nil {
-			// Log but don't fail - consistent with existing cleanup pattern
-			logDebug("MCP server session close error (non-fatal): %v\n", err)
-		}
-	}
-	if a.metricsServer != nil {
-		shutdownCtx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-		_ = a.metricsServer.Shutdown(shutdownCtx)
-		cancel()
-	}
-	if a.httpServer != nil {
-		ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-		defer cancel()
-		_ = a.httpServer.Shutdown(ctx)
-	}
-
-	return ctx, nil
-}
-
-func (tc *scenarioConfig) logDebug(format string, a ...any) {
-	if v, exists := tc.reqHeaders[server.TRANSACTION_ID_HEADER]; exists && v != "" {
-		format = fmt.Sprintf("(%s) %s", v, format)
-	}
-	fmt.Printf(format, a...)
-	getLogger().Printf(format, a...)
-}
-
-func (tc *scenarioConfig) logError(err error, withStack ...bool) error {
-	var sb = strings.Builder{}
-	sb.WriteString("Error")
-	if reqId, exists := tc.reqHeaders[server.TRANSACTION_ID_HEADER]; exists && reqId != "" {
-		fmt.Fprintf(&sb, " (%s)", reqId)
-	}
-	sb.WriteString(": ")
-	if len(withStack) > 0 && withStack[0] {
-		getLogger().Printf("%s%v\n%s\n", sb.String(), err, string(debug.Stack()))
-	} else {
-		getLogger().Printf("%s%v\n", sb.String(), err)
-	}
-	return fmt.Errorf("%s%v", sb.String(), err)
-}
-
-func (tc *scenarioConfig) saveValue(name, value string) {
-	tc.values[name] = value
-	tc.logDebug("Saved value %s: %s\n", name, value)
-}
-
-func (tc *scenarioConfig) theServiceIsRunning(ctx context.Context) error {
-	// Check that the server is actually running by sending a request to the health endpoint
-	for range 20 {
-		if err := tc.checkHealthEndpoint(); err != nil {
-			tc.logDebug("Error checking health endpoint: %v\n", err.Error())
-			time.Sleep(1 * time.Second)
-		} else {
-			return nil
-		}
-	}
-	return tc.logError(fmt.Errorf("service is not running"))
-}
-
-func (tc *scenarioConfig) thereAreSystemProviders(ctx context.Context) error {
-	if err := tc.iSendARequestImpl("GET", "/api/v1/evaluations/providers?scope=system&limit=100", "", "there are system providers"); err != nil {
-		return err
-	}
-	if tc.response.StatusCode != 200 {
-		return tc.logError(fmt.Errorf("expected 200 listing system providers, got %d: %s", tc.response.StatusCode, string(tc.body)))
-	}
-
-	var resp struct {
-		TotalCount int `json:"total_count"`
-	}
-	if err := json.Unmarshal(tc.body, &resp); err != nil {
-		return tc.logError(fmt.Errorf("failed to parse providers list: %w", err))
-	}
-
-	if resp.TotalCount == 0 {
-		tc.logDebug("Skipping scenario: no system providers found so skipping the scenario\n")
-		return godog.ErrSkip
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) thereAreSystemCollections(ctx context.Context) error {
-	if err := tc.iSendARequestImpl("GET", "/api/v1/evaluations/collections?scope=system&limit=100", "", "there are system collections"); err != nil {
-		return err
-	}
-	if tc.response.StatusCode != 200 {
-		return tc.logError(fmt.Errorf("expected 200 listing system collections, got %d: %s", tc.response.StatusCode, string(tc.body)))
-	}
-
-	var resp struct {
-		TotalCount int `json:"total_count"`
-		Items      []struct {
-			Resource struct {
-				ID string `json:"id"`
-			} `json:"resource"`
-			Name string `json:"name"`
-		} `json:"items"`
-	}
-	if err := json.Unmarshal(tc.body, &resp); err != nil {
-		return tc.logError(fmt.Errorf("failed to parse collections list: %w", err))
-	}
-
-	if resp.TotalCount == 0 {
-		tc.logDebug("Skipping scenario: no system collections found so skipping the scenario\n")
-		return godog.ErrSkip
-	}
-
-	// save the collection names for later use
-	for index, item := range resp.Items {
-		tc.saveValue(fmt.Sprintf("collection%d:id", index), item.Resource.ID)
-		tc.saveValue(fmt.Sprintf("collection%d:name", index), item.Name)
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) thereIsASystemCollectionWithId(ctx context.Context, id string) error {
-	if err := tc.iSendARequestImpl("GET", "/api/v1/evaluations/collections/"+id, "", "there is a system collection with id "+id); err != nil {
-		return err
-	}
-	if tc.response.StatusCode != 200 {
-		tc.logDebug("Skipping scenario: system collection with id %s not found\n", id)
-		return godog.ErrSkip
-	}
-
-	// save the collection id for later use
-	tc.saveValue("collection:id", id)
-	name, err := tc.getJsonPathValue("$.name")
-	if err != nil {
-		return err
-	}
-	nameStr, ok := name.(string)
-	if !ok {
-		return tc.logError(fmt.Errorf("expected name to be a string, got %T", name))
-	}
-	tc.saveValue("collection:name", nameStr)
-
-	return nil
-}
-
-func (tc *scenarioConfig) theValueIsSet(ctx context.Context, name string) error {
-	value, err := tc.getValue(name)
-	if err != nil {
-		return err
-	}
-	if strings.TrimSpace(value) == "" {
-		return tc.logError(fmt.Errorf("value %s is not set", name))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) checkHealthEndpoint() error {
-	if err := tc.iSendARequestImpl("GET", "/api/v1/health", "", "check health endpoint"); err != nil {
-		return tc.logError(fmt.Errorf("failed to send health check request: %w for URL %s", err, tc.apiFeature.baseURL.String()))
-	}
-	if tc.response.StatusCode != 200 {
-		return tc.logError(fmt.Errorf("expected status 200, got %d", tc.response.StatusCode))
-	}
-
-	match := "\"status\":\"healthy\""
-	if !strings.Contains(string(tc.body), match) {
-		return tc.logError(fmt.Errorf("expected body to contain %s, got %s", match, string(tc.body)))
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) iSetHeaderTo(paramName, paramValue string) error {
-	value, err := tc.getValue(paramValue)
-	if err != nil {
-		return err
-	}
-	tc.reqHeaders[paramName] = value
-	return nil
-}
-
-func (tc *scenarioConfig) iUnsetHeader(paramName string) error {
-	delete(tc.reqHeaders, paramName)
-	return nil
-}
-
-func (tc *scenarioConfig) iSetTransactionIdTo(paramValue string) error {
-	return tc.iSetHeaderTo(server.TRANSACTION_ID_HEADER, paramValue)
-}
-
-func (tc *scenarioConfig) iSendARequestTo(method, path string) error {
-	return tc.iSendARequestToWithBody(method, path, "")
-}
-
-func (tc *scenarioConfig) setDuration(dest *time.Duration, fieldName, paramValue string) error {
-	value, err := tc.getValue(paramValue)
-	if err != nil {
-		return err
-	}
-	*dest, err = time.ParseDuration(value)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to parse duration %q: %w", value, err))
-	}
-	if *dest <= 0 {
-		return tc.logError(fmt.Errorf("%s must be positive, got %q (%v)", fieldName, value, *dest))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) iSetWaitDeadlineTo(paramValue string) error {
-	return tc.setDuration(&tc.waitDeadline, "wait deadline", paramValue)
-}
-
-func (tc *scenarioConfig) iSetWaitIntervalTo(paramValue string) error {
-	return tc.setDuration(&tc.waitInterval, "wait interval", paramValue)
-}
-
-func (tc *scenarioConfig) iWaitForEvaluationJobStatus(expectedStatus string) error {
-	deadline := time.Now().Add(tc.waitDeadline)
-	var lastErr error
-	var lastStatus string
-	for time.Now().Before(deadline) {
-		if err := tc.iSendARequestImpl(http.MethodGet, "/api/v1/evaluations/jobs/{id}", "", "wait for evaluation job status"); err != nil {
-			lastErr = err
-			time.Sleep(tc.waitInterval)
-			continue
-		}
-		if tc.response != nil && tc.response.StatusCode == http.StatusOK {
-			status, err := tc.getJsonPath("$.status.state")
-			if status != "" {
-				lastStatus = status
-			}
-			if err != nil {
-				lastErr = err
-			} else if status == expectedStatus {
-				return nil
-			} else {
-				// Fail fast when the job has reached any terminal state other than the expected one.
-				if pkgapi.OverallState(status).IsTerminalState() {
-					// Get additional error context from the response for better diagnostics
-					message, _ := tc.getJsonPath("$.status.message.message")
-					if message != "" {
-						return tc.logError(fmt.Errorf("evaluation job reached terminal state %q (expected %q): %s", status, expectedStatus, message))
-					}
-					return tc.logError(fmt.Errorf("evaluation job reached terminal state %q (expected %q)", status, expectedStatus))
-				}
-				// we should not do this because it will be logged as an error
-				// lastErr = fmt.Errorf("expected status %q but got %q", expectedStatus, status)
-			}
-		} else if tc.response != nil {
-			lastErr = tc.logError(fmt.Errorf("unexpected response status %d", tc.response.StatusCode))
-		}
-		time.Sleep(tc.waitInterval)
-	}
-	if lastErr != nil {
-		return tc.logError(lastErr)
-	}
-	return tc.logError(fmt.Errorf("timed out after %v waiting for status %q, last status: %q", tc.waitDeadline, expectedStatus, lastStatus))
-}
-
-var errTestFileNotFound = errors.New("test file not found")
-
-// fvtDisconnected reports whether FVT is targeting a disconnected cluster
-// (ENVIRONMENT_ID contains "disconnected").
-func fvtDisconnected() bool {
-	return strings.Contains(strings.ToLower(os.Getenv("ENVIRONMENT_ID")), "disconnected")
-}
-
-// fvtBenchmarkTokenizer returns the expected benchmark tokenizer for FVT assertions and payloads.
-func fvtBenchmarkTokenizer() string {
-	if fvtDisconnected() {
-		return "/test_data/tokenizer"
-	}
-	return "google/flan-t5-small"
-}
-
-func (tc *scenarioConfig) findFile(fileName string) (string, error) {
-	file := filepath.Join(testDataRoot(), fileName)
-	_, err := os.Stat(file)
-	if err == nil {
-		return file, nil
-	}
-	if os.IsNotExist(err) {
-		return "", errTestFileNotFound
-	}
-	return "", tc.logError(fmt.Errorf("stat test file %s: %w", fileName, err))
-}
-
-func (tc *scenarioConfig) getFile(fileName string) (string, error) {
-	if jsonnetPath, err := tc.findFile(tc.jsonnetSiblingName(fileName)); err == nil {
-		return tc.evaluateJsonnetFile(jsonnetPath)
-	} else if !errors.Is(err, errTestFileNotFound) {
-		return "", err
-	}
-	filePath, err := tc.findFile(fileName)
-	if errors.Is(err, errTestFileNotFound) {
-		path, _ := os.Getwd()
-		return "", tc.logError(fmt.Errorf("test file %s not found in directory %s", fileName, path))
-	}
-	if err != nil {
-		return "", err
-	}
-	contents, err := os.ReadFile(filePath)
-	if err != nil {
-		return "", err
-	}
-	return string(contents), nil
-}
-
-func (tc *scenarioConfig) substituteValues(body string) (string, error) {
-	re := regexp.MustCompile(`\{\{([^}]*)\}\}`)
-	for strings.Contains(body, "{{") {
-		match := re.FindStringSubmatch(body)
-		if len(match) > 1 {
-			if after, ok := strings.CutPrefix(match[1], mlflowPrefix); ok {
-				// Use the literal after mlflow: as the experiment name. When MLflow is configured,
-				// it could be resolved from MLflow; for tests without MLflow, this allows name-based
-				// search to match stored jobs.
-				experimentName := after
-				if os.Getenv("MLFLOW_TRACKING_URI") == "" {
-					experimentName = ""
-				}
-				tc.logDebug("Substituting value '%s' with '%s'\n", match[1], experimentName)
-				body = strings.ReplaceAll(body, fmt.Sprintf("{{%s}}", match[1]), experimentName)
-			} else if raw, ok := strings.CutPrefix(match[1], envPrefix); ok {
-				envName, fallback, hasFallback := strings.Cut(raw, "|")
-				var value string
-				if v, found := gpuTestSuiteSubstValue(envName); found {
-					value = v
-				} else if envName == "FVT_BENCHMARK_TOKENIZER" {
-					value = fvtBenchmarkTokenizer()
-				} else if envValue, envOk := os.LookupEnv(envName); envOk {
-					value = envValue
-				} else if hasFallback {
-					value = fallback
-				} else {
-					value = ""
-				}
-				tc.logDebug("Substituting value '%s' with '%s'\n", match[1], value)
-				body = strings.ReplaceAll(body, fmt.Sprintf("{{%s}}", match[1]), value)
-			} else if after1, ok := strings.CutPrefix(match[1], valuePrefix); ok {
-				n := after1
-				v := tc.values[n]
-				tc.logDebug("Substituting value '%s' with '%s'\n", match[1], v)
-				body = strings.ReplaceAll(body, fmt.Sprintf("{{%s}}", match[1]), v)
-			} else {
-				return "", tc.logError(fmt.Errorf("unknown substitution value: %s", match[1]))
-			}
-		}
-	}
-	return body, nil
-}
-
-func (tc *scenarioConfig) getRequestBody(body string) (io.Reader, error) {
-	var err error
-	if body == "" {
-		return nil, nil
-	}
-	// this can be an inline body or a test file
-	if strings.HasPrefix(body, "file:/") {
-		// this returns the contents of the file as a string
-		body, err = tc.getFile(strings.TrimPrefix(body, "file:/"))
-		if err != nil {
-			return nil, err
-		}
-	}
-	// now do any substitution
-	body, err = tc.substituteValues(body)
-	if err != nil {
-		return nil, err
-	}
-	return strings.NewReader(body), nil
-}
-
-func (tc *scenarioConfig) addAsset(assetName, id string) {
-	//tc.assetsSync.Lock()
-	//defer tc.assetsSync.Unlock()
-	tc.assets[assetName] = append(tc.assets[assetName], id)
-	tc.logDebug("Added asset id %s for %s\n", id, assetName)
-}
-
-func (tc *scenarioConfig) removeAsset(assetName, id string) {
-	//tc.assetsSync.Lock()
-	//defer tc.assetsSync.Unlock()
-	ids := tc.assets[assetName]
-	if slices.Contains(ids, id) {
-		tc.assets[assetName] = slices.DeleteFunc(ids, func(s string) bool {
-			if s == id {
-				tc.logDebug("Removed asset id %s for %s\n", id, assetName)
-				return true
-			}
-			return false
-		})
-	}
-}
-
-func (tc *scenarioConfig) extractId(body []byte) (string, error) {
-	if len(body) > 0 {
-		obj := make(map[string]interface{})
-		err := json.Unmarshal(body, &obj)
-		if err != nil {
-			return "", tc.logError(fmt.Errorf("failed to unmarshal body %s: %w", string(body), err))
-		}
-		resource, ok := obj["resource"].(map[string]any)
-		if !ok {
-			return "", tc.logError(fmt.Errorf("response does not contain resource object: %s", string(body)))
-		}
-		id, ok := resource["id"].(string)
-		if !ok || id == "" {
-			return "", tc.logError(fmt.Errorf("response does not contain resource.id: %s", string(body)))
-		}
-		return id, nil
-	}
-	return "", nil
-}
-
-// pathDetails extracts the details from the path
-// the first match is the asset name
-// the second match is the asset type
-// the third match is the asset id
-// Handles: /api/v1/{name}, /api/v1/{name}/{asset}, /api/v1/{name}/{asset}/{id}
-// Uses [^/?]+ to stop at query strings
-var pathDetails = regexp.MustCompile(`^.*/api/v1/([^/?]+)(?:/([^/?]+))?(?:/([^/?]+))?.*$`)
-
-func (tc *scenarioConfig) getAssetDetails(path string) (string, string, string, error) {
-	if matches := pathDetails.FindStringSubmatch(path); len(matches) >= 4 {
-		return matches[1], matches[2], matches[3], nil
-	}
-	return "", "", "", tc.logError(fmt.Errorf("no first path segment found in path %s", path))
-}
-
-var valueExpression = regexp.MustCompile(`^(.*)[\s]*([+-])[\s]*(\d+)$`)
-
-func (tc *scenarioConfig) getValueExpression(id string) (string, int, error) {
-	matches := valueExpression.FindStringSubmatch(id)
-	if len(matches) >= 4 {
-		v, err := strconv.Atoi(matches[3])
-		if err != nil {
-			return "", 0, err
-		}
-		if matches[2] == "+" {
-			return strings.TrimRight(matches[1], " "), v, nil
-		}
-		return strings.TrimRight(matches[1], " "), -v, nil
-	}
-	return id, 0, nil
-}
-
-func (tc *scenarioConfig) getValue(id string) (string, error) {
-	// start with the full substitution
-	if value, err := tc.substituteValues(id); err == nil {
-		id = value
-	}
-	// Handle {variable} pattern by looking up in values map
-	if strings.HasPrefix(id, "{") && strings.HasSuffix(id, "}") {
-		n := strings.TrimSuffix(strings.TrimPrefix(id, "{"), "}")
-		v := tc.values[n]
-		if v == "" {
-			return "", tc.logError(fmt.Errorf("failed to find value for {%s}", n))
-		}
-		return v, nil
-	}
-	if strings.HasPrefix(id, valuePrefix) {
-		n := strings.TrimPrefix(id, valuePrefix)
-		v := tc.values[n]
-		if v == "" {
-			return "", tc.logError(fmt.Errorf("failed to find value %s", n))
-		}
-		return v, nil
-	}
-	return id, nil
-}
-
-func (tc *scenarioConfig) getEndpoint(path string) (string, error) {
-	check := true
-	for check {
-		if strings.Contains(path, fmt.Sprintf("{{%s", valuePrefix)) {
-			re := regexp.MustCompile(`\{\{([^}]*)\}\}`)
-			match := re.FindStringSubmatch(path)
-			if len(match) > 1 {
-				v, err := tc.getValue(match[1])
-				if err != nil {
-					return "", tc.logError(fmt.Errorf("failed to substitute value: %s", err.Error()))
-				}
-				path = strings.ReplaceAll(path, fmt.Sprintf("{{%s}}", match[1]), v)
-			} else {
-				// no more matches found
-				check = false
-			}
-		} else {
-			check = false
-		}
-	}
-
-	if strings.Contains(path, "{id}") {
-		if tc.lastId == "" {
-			return "", tc.logError(fmt.Errorf("last ID is not set"))
-		}
-		path = strings.Replace(path, "{id}", tc.lastId, 1)
-	}
-
-	if isMetricsScrapePath(path) {
-		if tc.apiFeature.metricsBaseURL == nil {
-			return "", tc.logError(fmt.Errorf(
-				"METRICS_URL is required when SERVER_URL is set (metrics are served on a separate port, not through kube-rbac-proxy)",
-			))
-		}
-		return joinBaseURL(tc.apiFeature.metricsBaseURL, path), nil
-	}
-
-	endpoint := path
-	if !strings.HasPrefix(endpoint, tc.apiFeature.baseURL.String()) {
-		endpoint = joinBaseURL(tc.apiFeature.baseURL, path)
-	}
-
-	return endpoint, nil
-}
-
-func (tc *scenarioConfig) iSendARequestToWithInlineBody(method, path string, body *godog.DocString) error {
-	if body == nil {
-		return tc.logError(fmt.Errorf("inline body is missing"))
-	}
-	return tc.iSendARequestToWithBody(method, path, body.Content)
-}
-
-func (tc *scenarioConfig) iSendARequestToWithBody(method, path, body string) error {
-	return tc.iSendARequestImpl(method, path, body, "")
-}
-
-func (tc *scenarioConfig) iSendARequestImpl(method, path, body, caller string) error {
-	endpoint, err := tc.getEndpoint(path)
-	if err != nil {
-		return err
-	}
-	tc.lastURL = endpoint
-	tc.lastMethod = method
-	entity, err := tc.getRequestBody(body)
-	if err != nil {
-		return err
-	}
-	if caller != "" {
-		tc.logDebug("Sending %s request to %s by %s with body %s\n", method, endpoint, caller, body)
-	} else {
-		tc.logDebug("Sending %s request to %s with body %s\n", method, endpoint, body)
-	}
-	req, err := http.NewRequest(method, endpoint, entity)
-	if err != nil {
-		tc.logDebug("Failed to create request: %v\n", err)
-		return err
-	}
-	scrapeMetrics := isMetricsScrapePath(path)
-	if !scrapeMetrics {
-		if authToken := os.Getenv("AUTH_TOKEN"); authToken != "" {
-			req.Header.Set("Authorization", "Bearer "+authToken)
-		}
-		if tenant := os.Getenv("X_TENANT"); tenant != "" {
-			req.Header.Set("X-Tenant", tenant)
-		}
-	}
-
-	for k, v := range tc.reqHeaders {
-		req.Header.Set(k, v)
-	}
-
-	tc.response, err = tc.apiFeature.client.Do(req)
-	if err != nil {
-		tc.logDebug("Failed to send request: %v\n", err)
-		return err
-	}
-
-	defer func() {
-		// we do this for now as request ids are supposed to be unique per request
-		_ = tc.iUnsetHeader(server.TRANSACTION_ID_HEADER)
-	}()
-
-	tc.body, err = io.ReadAll(tc.response.Body)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = tc.response.Body.Close() }()
-
-	if len(tc.body) > 0 && len(tc.body) < 1024*5 {
-		tc.logDebug("Response status %d for %s %s with body %s\n", tc.response.StatusCode, method, endpoint, string(tc.body))
-	} else {
-		tc.logDebug("Response status %d for %s %s\n", tc.response.StatusCode, method, endpoint)
-	}
-
-	// capture resource id for create (evaluation job or collection)
-	if method == http.MethodPost && (tc.response.StatusCode == http.StatusAccepted || tc.response.StatusCode == http.StatusCreated) {
-		_, assetName, _, err := tc.getAssetDetails(endpoint)
-		if err != nil {
-			return err
-		}
-		if assetName != "" {
-			tc.lastId, err = tc.extractId(tc.body)
-			if err != nil {
-				return err
-			}
-			if tc.lastId == "" {
-				return tc.logError(fmt.Errorf("response does not contain an ID in response %s", string(tc.body)))
-			}
-			tc.addAsset(assetName, tc.lastId)
-			tc.values["id"] = tc.lastId
-		}
-	}
-
-	if method == http.MethodDelete {
-		_, assetName, _, err := tc.getAssetDetails(endpoint)
-		if err != nil {
-			return err
-		}
-		if assetName != "" {
-			_, _, id, err := tc.getAssetDetails(endpoint)
-			if err != nil {
-				return err
-			}
-			if id == "" {
-				return tc.logError(fmt.Errorf("no ID found in path %s", endpoint))
-			}
-			parsedURL, err := url.Parse(endpoint)
-			if err != nil {
-				return tc.logError(fmt.Errorf("failed to parse endpoint %s: %w", endpoint, err))
-			}
-			if parsedURL.Query().Get("hard_delete") == "true" {
-				tc.removeAsset(assetName, id)
-			}
-		}
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseStatusShouldBe(status int) error {
-	if tc.response.StatusCode != status {
-		return tc.logError(fmt.Errorf("expected status %d, got %d for request %s %s with response %s", status, tc.response.StatusCode, tc.lastMethod, tc.lastURL, string(tc.body)))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseStatusShouldBeOr(status1, status2 int) error {
-	if (tc.response.StatusCode != status1) && (tc.response.StatusCode != status2) {
-		return tc.logError(fmt.Errorf("expected status %d or %d, got %d for request %s %s with response %s", status1, status2, tc.response.StatusCode, tc.lastMethod, tc.lastURL, string(tc.body)))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseContentTypeShouldBe(contentType string) error {
-	expected, err := tc.getValue(contentType)
-	if err != nil {
-		return err
-	}
-	actual := tc.response.Header.Get("Content-Type")
-	if !strings.HasPrefix(actual, expected) {
-		return tc.logError(fmt.Errorf("expected Content-Type to start with %q, got %q for request %s %s", expected, actual, tc.lastMethod, tc.lastURL))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseBodyShouldContain(text string) error {
-	expected, err := tc.getValue(text)
-	if err != nil {
-		return err
-	}
-	body := string(tc.body)
-	if !strings.Contains(body, expected) {
-		return tc.logError(fmt.Errorf("expected response body to contain %q for request %s %s, got %q", expected, tc.lastMethod, tc.lastURL, body))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseShouldContainWithValue(key, value string) error {
-	var data map[string]interface{}
-	if err := json.Unmarshal(tc.body, &data); err != nil {
-		return tc.logError(err)
-	}
-
-	v, err := tc.getValue(value)
-	if err != nil {
-		return err
-	}
-
-	if data[key] != v {
-		return tc.logError(fmt.Errorf("expected %s to be %s, got %v in %s", key, v, data[key], asPrettyJson(string(tc.body))))
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseShouldContain(key string) error {
-	var data map[string]interface{}
-	if err := json.Unmarshal(tc.body, &data); err != nil {
-		return tc.logError(err)
-	}
-
-	k, err := tc.getValue(key)
-	if err != nil {
-		return err
-	}
-
-	if _, ok := data[k]; !ok {
-		return tc.logError(fmt.Errorf("response does not contain key: %s in %s", k, asPrettyJson(string(tc.body))))
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseShouldContainPrometheusMetrics() error {
-	bodyStr := string(tc.body)
-	if !strings.Contains(bodyStr, "# HELP") || !strings.Contains(bodyStr, "# TYPE") {
-		return tc.logError(fmt.Errorf("response does not appear to be Prometheus metrics format"))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseShouldBeJSON() error {
-	var data interface{}
-	if err := json.Unmarshal(tc.body, &data); err != nil {
-		return tc.logError(err)
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theMetricsShouldInclude(metricName string) error {
-	bodyStr := string(tc.body)
-	if !strings.Contains(bodyStr, metricName) {
-		return tc.logError(fmt.Errorf("metrics do not include %s", metricName))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theMetricsShouldShowRequestCountFor(path string) error {
-	bodyStr := string(tc.body)
-	// Check if metrics contain the path
-	if !strings.Contains(bodyStr, path) {
-		return tc.logError(fmt.Errorf("metrics do not show requests for path %s", path))
-	}
-	return nil
-}
-
-func asPrettyJson(s string) string {
-	js := make(map[string]interface{})
-	err := json.Unmarshal([]byte(s), &js)
-	if err != nil {
-		return s
-	}
-	ns, err := json.MarshalIndent(js, "", "  ")
-	if err != nil {
-		return s
-	}
-	return string(ns)
-}
-
-func (tc *scenarioConfig) compareJSONSchema(expectedSchema string, actualResponse string) error {
-	expectedSchemaLoader := gojsonschema.NewStringLoader(expectedSchema)
-	return tc.validateJSONSchema(expectedSchemaLoader, actualResponse)
-}
-
-func (tc *scenarioConfig) compareJSONSchemaFile(schemaFile string, actualResponse string) error {
-	schemaContent, err := tc.getFile(schemaFile)
-	if err != nil {
-		return tc.logError(fmt.Errorf("schema file %s: %w", schemaFile, err))
-	}
-	return tc.compareJSONSchema(schemaContent, actualResponse)
-}
-
-func (tc *scenarioConfig) validateJSONSchema(expectedSchemaLoader gojsonschema.JSONLoader, actualResponse string) error {
-	actualResultLoader := gojsonschema.NewStringLoader(actualResponse)
-	result, validateErr := gojsonschema.Validate(expectedSchemaLoader, actualResultLoader)
-	if validateErr != nil {
-		fmt.Printf("The actual response %s does not match expected schema with error:\n", asPrettyJson(actualResponse))
-		if result != nil {
-			for _, err := range result.Errors() {
-				fmt.Printf("- %s value = %s\n", err, err.Value())
-			}
-		}
-		fmt.Printf("- error %s\n", validateErr.Error())
-		return validateErr
-	}
-	if len(result.Errors()) > 0 {
-		fmt.Printf("The actual response %s does not match expected schema with error:\n", asPrettyJson(actualResponse))
-		for _, err := range result.Errors() {
-			fmt.Printf("- %s value = %s\n", err, err.Value())
-		}
-		return fmt.Errorf("the response does not match the expected JSON schema")
-	}
-	if result.Valid() {
-		return nil
-	}
-	return fmt.Errorf("failed to validate the response %s but no error detected", asPrettyJson(actualResponse))
-}
-
-func (tc *scenarioConfig) theResponseShouldHaveSchemaAs(body *godog.DocString) error {
-	return tc.compareJSONSchema(body.Content, string(tc.body))
-}
-
-func (tc *scenarioConfig) theResponseShouldHaveSchemaFromFile(filePath string) error {
-	filePath = strings.TrimPrefix(filePath, "file:/")
-	return tc.compareJSONSchemaFile(filePath, string(tc.body))
-}
-
-func (tc *scenarioConfig) unquoteJsonPath(jsonPath string) string {
-	s := strings.ReplaceAll(jsonPath, "&quot;", "\"")
-	// s = strings.ReplaceAll(jsonPath, "&#39;", "'")
-	return s
-}
-
-func (tc *scenarioConfig) getJsonPath(jsonPath string) (string, error) {
-	jsonPath = tc.unquoteJsonPath(jsonPath)
-
-	// first check the jsonpath is valid
-	_, err := jsonpath.New(jsonPath)
-	if err != nil {
-		return "", fmt.Errorf("failed to validate JSON path %s: %w : %s", jsonPath, err, asPrettyJson(string(tc.body))) // logging of the error is done by the caller
-	}
-
-	raw, err := tc.getJsonPathValue(jsonPath)
-	if err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%v", raw), nil
-}
-
-func (tc *scenarioConfig) getJsonPathValue(jsonPath string) (interface{}, error) {
-	var respMap map[string]interface{}
-	err := json.Unmarshal(tc.body, &respMap)
-	if err != nil {
-		return "", err // logging of the error is done by the caller
-	}
-	path := jsonPath
-	if !strings.HasPrefix(path, "$") {
-		path = "$." + path
-	}
-	foundValue, err := jsonpath.Get(path, respMap)
-	if err != nil {
-		return "", fmt.Errorf("failed to get JSON path %s in %s: %w", jsonPath, asPrettyJson(string(tc.body)), err) // logging of the error is done by the caller
-	}
-	return foundValue, nil
-}
-
-func (tc *scenarioConfig) theResponseShouldContainAtJSONPath(expectedValue string, jsonPath string) error {
-	_, _, err := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, "contains")
-	return err
-}
-
-func (tc *scenarioConfig) theResponseShouldEqualAtJSONPath(expectedValue string, jsonPath string) error {
-	_, _, err := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, "==")
-	return err
-}
-
-func (tc *scenarioConfig) theResponseShouldContainAtJSONPathAtLeast(expectedValue string, jsonPath string) error {
-	_, _, err := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, ">=")
-	return err
-}
-
-func (tc *scenarioConfig) theResponseShouldMatchAtJSONPath(expectedValue string, jsonPath string) error {
-	_, _, err := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, "matches")
-	return err
-}
-
-func (tc *scenarioConfig) theResponseShouldContainAtJSONPathImpl(expectedValue string, jsonPath string, match string) (bool, string, error) {
-	expanded, err := tc.substituteValues(expectedValue)
-	if err != nil {
-		return false, "", err
-	}
-	expectedValue = expanded
-
-	foundValue, err := tc.getJsonPath(jsonPath)
-	if err != nil {
-		// true because the path is not found
-		return true, foundValue, tc.logError(err)
-	}
-
-	if rawExpr, ok := strings.CutPrefix(expectedValue, regexpPrefix); ok {
-		expr, err := regexp.Compile(rawExpr)
-		if err != nil {
-			return false, foundValue, tc.logError(fmt.Errorf("invalid regex %q: %w", rawExpr, err))
-		}
-		if expr.MatchString(foundValue) {
-			tc.logDebug("Value %s matches regex %s in path %s", foundValue, rawExpr, jsonPath)
-			return false, foundValue, nil
-		}
-	}
-
-	values := strings.SplitSeq(expectedValue, "|")
-	for value := range values {
-		switch match {
-		case "==", "equals":
-			// first try an exact string match
-			if foundValue == strings.TrimSpace(value) {
-				return false, foundValue, nil
-			}
-			// then try a float match
-			if fv, err := strconv.ParseFloat(foundValue, 64); err == nil {
-				if ex, err := strconv.ParseFloat(strings.TrimSpace(value), 64); err == nil {
-					// compare the floats to 15 decimal places
-					if math.Abs(fv-ex) < 0.0000000000000001 {
-						return false, foundValue, nil
-					}
-				}
-			}
-		case "<=":
-			fv, err := strconv.ParseFloat(foundValue, 64)
-			if err != nil {
-				return false, foundValue, tc.logError(fmt.Errorf("failed to parse found value %s as float: %w", foundValue, err))
-			}
-			ex, err := strconv.ParseFloat(strings.TrimSpace(value), 64)
-			if err != nil {
-				return false, foundValue, tc.logError(fmt.Errorf("failed to parse expected value %s as float: %w", value, err))
-			}
-			if fv <= ex {
-				return false, foundValue, nil
-			}
-		case ">=":
-			fv, err := strconv.ParseFloat(foundValue, 64)
-			if err != nil {
-				return false, foundValue, tc.logError(fmt.Errorf("failed to parse found value %s as float: %w", foundValue, err))
-			}
-			ex, err := strconv.ParseFloat(strings.TrimSpace(value), 64)
-			if err != nil {
-				return false, foundValue, tc.logError(fmt.Errorf("failed to parse expected value %s as float: %w", value, err))
-			}
-			if fv >= ex {
-				return false, foundValue, nil
-			}
-		case "contains":
-			if strings.Contains(foundValue, strings.TrimSpace(value)) {
-				return false, foundValue, nil
-			}
-		case "matches":
-			expr, err := regexp.Compile(value)
-			if err != nil {
-				return false, foundValue, tc.logError(fmt.Errorf("invalid regex %q: %w", strings.TrimSpace(value), err))
-			}
-			if expr.MatchString(foundValue) {
-				return false, foundValue, nil
-			}
-		}
-	}
-
-	return true, foundValue, tc.logError(fmt.Errorf("expected %s to be %s but was %s in %s", jsonPath, expectedValue, foundValue, asPrettyJson(string(tc.body))))
-}
-
-func (tc *scenarioConfig) theResponseShouldNotContainAtJSONPath(expectedValue string, jsonPath string) error {
-	_, found, _ := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, "contains")
-	if strings.Contains(strings.TrimSpace(found), strings.TrimSpace(expectedValue)) {
-		return tc.logError(fmt.Errorf("expected %s to not contain %s but found %s in %s", jsonPath, expectedValue, found, asPrettyJson(string(tc.body))))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theResponseShouldNotEqualAtJSONPath(expectedValue string, jsonPath string) error {
-	_, found, _ := tc.theResponseShouldContainAtJSONPathImpl(expectedValue, jsonPath, "==")
-	if strings.TrimSpace(found) == strings.TrimSpace(expectedValue) {
-		return tc.logError(fmt.Errorf("expected %s to not equal %s but found %s in %s", jsonPath, expectedValue, found, asPrettyJson(string(tc.body))))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theArrayAtPathInResponseShouldHaveLength(jsonPath string, lengthStr string) error {
-	value, add, err := tc.getValueExpression(lengthStr)
-	if err != nil {
-		return err
-	}
-	value, err = tc.getValue(value)
-	if err != nil {
-		return tc.logError(err)
-	}
-	length, err := strconv.Atoi(value)
-	if err != nil {
-		return tc.logError(fmt.Errorf("expected integer length, got %q: %w", value, err))
-	}
-	length += add
-	raw, err := tc.getJsonPathValue(jsonPath)
-	if err != nil {
-		return err
-	}
-	arr, ok := raw.([]any)
-	if !ok {
-		return tc.logError(fmt.Errorf("value at path %s is not an array, got %T", jsonPath, raw))
-	}
-	if len(arr) != length {
-		return tc.logError(fmt.Errorf("expected array at path %s to have length %d, got %d in %s", jsonPath, length, len(arr), asPrettyJson(string(tc.body))))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theArrayAtPathInResponseShouldHaveLengthAtLeast(jsonPath string, minLengthStr string) error {
-	value, add, err := tc.getValueExpression(minLengthStr)
-	if err != nil {
-		return err
-	}
-	value, err = tc.getValue(value)
-	if err != nil {
-		return err
-	}
-	minLength, err := strconv.Atoi(value)
-	if err != nil {
-		return tc.logError(fmt.Errorf("expected integer min length, got %q: %w", value, err))
-	}
-	minLength += add
-	raw, err := tc.getJsonPathValue(jsonPath)
-	if err != nil {
-		return err
-	}
-	arr, ok := raw.([]interface{})
-	if !ok {
-		return tc.logError(fmt.Errorf("value at path %s is not an array, got %T", jsonPath, raw))
-	}
-	if len(arr) < minLength {
-		return tc.logError(fmt.Errorf("expected array at path %s to have length >= %d, got %d in %s", jsonPath, minLength, len(arr), asPrettyJson(string(tc.body))))
-	}
-	return nil
-}
-
-func getJsonPointer(path string) string {
-	// Strip JSONPath root indicators: $. or $
-	path = strings.TrimPrefix(path, "$.")
-	path = strings.TrimPrefix(path, "$")
-
-	// Ensure it starts with / for JSON Pointer spec
-	if !strings.HasPrefix(path, "/") {
-		path = "/" + path
-	}
-	// Convert dot notation to slash notation
-	path = strings.ReplaceAll(path, ".", "/")
-
-	// Convert array notation [N] to /N/ for JSON Pointer spec
-	// e.g., /benchmarks[0]/id becomes /benchmarks/0/id
-	re := regexp.MustCompile(`\[(\d+)\]`)
-	path = re.ReplaceAllString(path, "/$1")
-
-	return path
-}
-
-func (tc *scenarioConfig) theFieldShouldBeSaved(path string, name string) error {
-	jsonParsed, err := gabs.ParseJSON(tc.body)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to parse JSON response: %w", err))
-	}
-	// This directly uses a JSON pointer path
-	pathObj, err := jsonParsed.JSONPointer(getJsonPointer(path))
-	if err != nil {
-		return tc.logError(fmt.Errorf("path %v does not exist in \n%s", path, string(tc.body)))
-	}
-	finalResult, ok := pathObj.Data().(string)
-	if !ok {
-		if floatResult, ok := pathObj.Data().(float64); ok {
-			finalResult = strconv.FormatFloat(floatResult, 'f', -1, 64)
-		} else {
-			return tc.logError(fmt.Errorf("expected %s to be a string or float64 but got %T", path, pathObj.Data()))
-		}
-	}
-	if strings.HasPrefix(name, valuePrefix) {
-		realName := strings.TrimPrefix(name, valuePrefix)
-		tc.saveValue(realName, finalResult)
-		tc.logDebug("Saved value %s as %s\n", realName, finalResult)
-	} else {
-		return tc.logError(fmt.Errorf("unexpected value %s, should start with '%s'", name, valuePrefix))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) fixThisStep() error {
-	tc.logDebug("TODO: fix this step")
-	return godog.ErrSkip
-}
-
-func (tc *scenarioConfig) requireMetricsURLForRemoteServer(ctx context.Context, sc *godog.Scenario) (context.Context, error) {
-	// Not a @metrics scenario.
-	// Input: any scenario without the @metrics tag.
-	// Behavior: no-op; other features are unaffected by METRICS_URL requirements.
-	if !scenarioHasTag(sc, "metrics") {
-		return ctx, nil
-	}
-
-	// Local embedded-server mode.
-	// Input: SERVER_URL unset; METRICS_URL optional (defaults via resolveMetricsBaseURL).
-	// Behavior: run the scenario; /metrics is served on the main API router in local mode.
-	if strings.TrimSpace(os.Getenv("SERVER_URL")) == "" {
-		return ctx, nil
-	}
-
-	// Remote/cluster mode with METRICS_URL configured.
-	// Input: SERVER_URL=https://evalhub.example.com, METRICS_URL=http://evalhub-metrics.<ns>.svc:8081.
-	// Behavior: run the scenario; scrape requests use the dedicated metrics port.
-	if strings.TrimSpace(os.Getenv(envMetricsURL)) != "" {
-		return ctx, nil
-	}
-
-	// Remote/cluster mode without METRICS_URL.
-	// Input: SERVER_URL set, METRICS_URL unset.
-	// Behavior: skip the scenario (scraping via SERVER_URL would hit kube-rbac-proxy and fail with 403).
-	tc.logDebug(
-		"Skipping scenario: METRICS_URL is required when SERVER_URL is set (metrics are served on a separate port, not through kube-rbac-proxy)\n",
-	)
-	return ctx, godog.ErrSkip
-}
-
-func (tc *scenarioConfig) saveScenarioName(ctx context.Context, sc *godog.Scenario) (context.Context, error) {
-	tc.scenarioName = sc.Name
-	tc.jsonnetQueueEnabled = nil
-	return ctx, nil
-}
-
-func (tc *scenarioConfig) queueIsEnabledForJsonnetPayloads() error {
-	queueOn := true
-	tc.jsonnetQueueEnabled = &queueOn
-	logDebug("Queue enabled for jsonnet payloads\n")
-	return nil
-}
-
-func (tc *scenarioConfig) assetCleanup(ctx context.Context, sc *godog.Scenario, err error) (context.Context, error) {
-	for assetName, ids := range tc.assets {
-		clonedIDs := slices.Clone(ids)
-		hardDelete := false
-		url := assetName
-		switch assetName {
-		case "evaluations":
-			url = "evaluations/jobs"
-			hardDelete = true
-		case "jobs":
-			url = "evaluations/jobs"
-			hardDelete = true
-		case "collections":
-			url = "evaluations/collections"
-		case "providers":
-			url = "evaluations/providers"
-		}
-		for _, id := range clonedIDs {
-			var path string
-			if hardDelete {
-				path = fmt.Sprintf("/api/v1/%s/%s?hard_delete=true", url, id)
-			} else {
-				path = fmt.Sprintf("/api/v1/%s/%s", url, id)
-			}
-			err := tc.iSendARequestImpl("DELETE", path, "", "asset cleanup")
-			if err != nil {
-				return ctx, tc.logError(fmt.Errorf("failed to delete asset %s with id '%s': %w", assetName, id, err))
-			}
-			err = tc.theResponseStatusShouldBe(204)
-			if err != nil {
-				_ = tc.logError(fmt.Errorf("failed to delete asset %s expected status %d but got %d: %w", tc.lastURL, 204, tc.response.StatusCode, err))
-				// return ctx, err
-			} else {
-				tc.logDebug("Deleted asset %s with status %d\n", path, tc.response.StatusCode)
-			}
-		}
-	}
-	tc.assets = nil
-	return ctx, nil
-}
-
-func createScenarioConfig(apiConfig *apiFeature) *scenarioConfig {
-	conf := new(scenarioConfig)
-	conf.reqHeaders = make(map[string]string)
-	conf.assets = make(map[string][]string)
-	conf.values = make(map[string]string)
-	conf.apiFeature = apiConfig
-
-	conf.waitDeadline = 30 * time.Minute
-	conf.waitInterval = 1 * time.Minute
-
-	return conf
-}
-
-func setUpTestConf() {
-	apiFeature, err := createApiFeature()
-	if err != nil {
-		panic(logError(fmt.Errorf("failed to create API feature: %v", err)))
-	}
-	apiFeat = apiFeature
-}
-
-func waitForService() {
-	tc := createScenarioConfig(apiFeat)
-	if err := tc.theServiceIsRunning(context.Background()); err != nil {
-		panic("Stopped API Tests. Service is not ready for testing.\n")
-	}
-}
-
-func tidyUpTests() {
-	if apiFeat != nil {
-		_, _ = apiFeat.cleanup(context.Background(), nil, nil)
-	}
-	if s, ok := logger.Writer().(*os.File); ok {
-		err := s.Close()
-		if err != nil {
-			panic(fmt.Sprintf("Failed to close logger file: %v\n", err))
-		}
-	}
-}
-
-func checkModelEndpoint() {
-	modelEndpointConnectivity = modelEndpointUnchecked
-
-	modelURL := os.Getenv("MODEL_URL")
-	if modelURL == "" {
-		logDebug("MODEL_URL not set, skipping model endpoint pre-flight check\n")
-		return
-	}
-
-	fmt.Printf("Checking model endpoint connectivity: %s\n", modelURL)
-
-	client := &http.Client{
-		Timeout: 10 * time.Second,
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				MinVersion:         tls.VersionTLS12,
-				InsecureSkipVerify: true, //nolint:gosec
-			},
-		},
-	}
-
-	maxRetries := 3
-	numRetries := 0
-	shouldRetry := func() bool { return numRetries < maxRetries }
-	notReadyStatus := func(statusCode int) bool { return statusCode == 503 }
-	retryDelay := 10 * time.Second
-
-	for shouldRetry() {
-		resp, err := client.Get(modelURL) //nolint:gosec // This is a test, we don't need to be too strict about the HTTP client
-		if err != nil {
-			var dnsErr *net.DNSError
-			if errors.As(err, &dnsErr) {
-				logDebug("WARNING: Cannot resolve model endpoint DNS for %s (test runner may be outside the cluster), proceeding with tests\n", modelURL)
-				return
-			}
-			logDebug("WARNING: Model endpoint %s is not reachable: %v\n", modelURL, err)
-			logDebug("Evaluation job scenarios will be skipped.\n")
-			modelEndpointConnectivity = modelEndpointUnreachable
-			return
-		}
-		status := resp.StatusCode
-		_, _ = io.Copy(io.Discard, resp.Body)
-		_ = resp.Body.Close()
-		logDebug("Model endpoint preflight GET %s returned HTTP %d\n", modelURL, status)
-
-		// 401 on the OpenAI-compatible base URL (/v1) still means the server is up;
-		// unauthenticated GET often cannot list models. Do not treat 404 as reachable
-		// (that can mask a bad MODEL_URL).
-		reachableWithoutAuth := status == http.StatusUnauthorized
-
-		if numRetries < maxRetries-1 && notReadyStatus(status) {
-			logDebug("WARNING: Model endpoint %s is not ready (HTTP %d), waiting %s before retrying\n", modelURL, status, retryDelay)
-			time.Sleep(retryDelay)
-		} else if status >= 200 && status < 300 {
-			logDebug("Model endpoint %s is reachable (HTTP %d)\n", modelURL, status)
-			modelEndpointConnectivity = modelEndpointReachable
-			return
-		} else if reachableWithoutAuth {
-			logDebug("Model endpoint %s is reachable (HTTP %d; auth expected for /v1 base URL)\n", modelURL, status)
-			modelEndpointConnectivity = modelEndpointReachable
-			return
-		} else {
-			logDebug("WARNING: Model endpoint %s returned HTTP %d, treating as unreachable\n", modelURL, status)
-			logDebug("Evaluation job scenarios will be skipped.\n")
-			modelEndpointConnectivity = modelEndpointUnreachable
-			return
-		}
-		numRetries++
-	}
-}
-
-func (tc *scenarioConfig) theModelEndpointIsReachable() error {
-	switch modelEndpointConnectivity {
-	case modelEndpointUnreachable:
-		logDebug("Model endpoint is not reachable, skipping evaluation job scenario %s\n", tc.scenarioName)
-		return godog.ErrSkip
-	case modelEndpointUnchecked, modelEndpointReachable:
-		return nil
-	default:
-		return nil
-	}
-}
-
-// A bit of a hack to have some checks that the regexes are working as expected
-func checkRegexes() {
-	tc := createScenarioConfig(apiFeat)
-	paths := [][]string{
-		{"/api/v1/evaluations", "evaluations", "", ""},
-		{"/api/v1/evaluations/jobs", "evaluations", "jobs", ""},
-		{"/api/v1/evaluations/jobs/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "jobs", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"/api/v1/evaluations/jobs/f02b16a2-1990-4626-b24d-1cff3febdbfb/update", "evaluations", "jobs", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"/api/v1/evaluations/collections", "evaluations", "collections", ""},
-		{"/api/v1/evaluations/collections/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "collections", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"/api/v1/evaluations/providers", "evaluations", "providers", ""},
-		{"/api/v1/evaluations/providers/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "providers", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"http://localhost:8080/api/v1/evaluations", "evaluations", "", ""},
-		{"http://localhost:8080/api/v1/evaluations/jobs", "evaluations", "jobs", ""},
-		{"http://localhost:8080/api/v1/evaluations/jobs/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "jobs", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"http://localhost:8080/api/v1/evaluations/jobs/f02b16a2-1990-4626-b24d-1cff3febdbfb/update", "evaluations", "jobs", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"http://localhost:8080/api/v1/evaluations/collections", "evaluations", "collections", ""},
-		{"http://localhost:8080/api/v1/evaluations/collections/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "collections", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"http://localhost:8080/api/v1/evaluations/providers", "evaluations", "providers", ""},
-		{"http://localhost:8080/api/v1/evaluations/providers/f02b16a2-1990-4626-b24d-1cff3febdbfb", "evaluations", "providers", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-		{"http://localhost:8080/api/v1/evaluations/providers?a=b", "evaluations", "providers", ""},
-		{"http://localhost:8080/api/v1/evaluations/providers/f02b16a2-1990-4626-b24d-1cff3febdbfb?a=b", "evaluations", "providers", "f02b16a2-1990-4626-b24d-1cff3febdbfb"},
-	}
-	for _, path := range paths {
-		name, asset, id, err := tc.getAssetDetails(path[0])
-		if err != nil {
-			panic(tc.logError(fmt.Errorf("failed to parse details from path %s: %v", path, err)))
-		}
-		if name != path[1] {
-			panic(tc.logError(fmt.Errorf("expected asset name %s for path %s, got %s", path[1], path[0], name)))
-		}
-		if asset != path[2] {
-			panic(tc.logError(fmt.Errorf("expected asset %s for path %s, got %s", path[2], path[0], asset)))
-		}
-		if id != path[3] {
-			panic(tc.logError(fmt.Errorf("expected asset id %s for path %s, got %s", path[3], path[0], id)))
-		}
-	}
-
-	values := [][]string{
-		{"{{value:num_providers}}+2", "{{value:num_providers}}", "2"},
-		{"{{value:num_providers}} + 2", "{{value:num_providers}}", "2"},
-		{"{{value:num_providers}}-2", "{{value:num_providers}}", "-2"},
-		{"{{value:num_providers}} - 2", "{{value:num_providers}}", "-2"},
-	}
-	for _, value := range values {
-		v, count, err := tc.getValueExpression(value[0])
-		if err != nil {
-			panic(tc.logError(fmt.Errorf("failed to parse value expression %s: %v", value[0], err)))
-		}
-		if v != value[1] {
-			panic(tc.logError(fmt.Errorf("expected value '%s' for value expression '%s', got '%s'", value[1], value[0], v)))
-		}
-		if fmt.Sprintf("%d", count) != value[2] {
-			panic(tc.logError(fmt.Errorf("expected count %s for value expression %s, got %d", value[1], value[0], count)))
-		}
-	}
-}
-
-func InitializeTestSuite(ctx *godog.TestSuiteContext) {
-	http.DefaultTransport.(*http.Transport).TLSClientConfig = &tls.Config{
-		MinVersion: tls.VersionTLS12,
-		MaxVersion: tls.VersionTLS13,
-		//nolint:gosec
-		InsecureSkipVerify: true,
-	}
-
-	if authToken := os.Getenv("AUTH_TOKEN"); authToken != "" {
-		logDebug("Using Authorization header with token\n")
-	}
-	if tenant := os.Getenv("X_TENANT"); tenant != "" {
-		logDebug("Using X-Tenant header with value %s\n", tenant)
-	}
-	if metricsURL := os.Getenv(envMetricsURL); metricsURL != "" {
-		logDebug("Using METRICS_URL for Prometheus scrape requests: %s\n", metricsURL)
-	}
-
-	ctx.BeforeSuite(checkRegexes)
-
-	ctx.BeforeSuite(setUpTestConf)
-	ctx.BeforeSuite(waitForService)
-	ctx.BeforeSuite(checkModelEndpoint)
-
-	// Initialize GPU test suite hooks
-	InitializeGPUTestSuite(ctx)
-
-	ctx.AfterSuite(tidyUpTests)
-}
-
-func InitializeScenario(ctx *godog.ScenarioContext) {
-	tc := createScenarioConfig(apiFeat)
-
-	ctx.Before(tc.saveScenarioName)
-	ctx.Before(tc.requireMetricsURLForRemoteServer)
-	ctx.After(tc.assetCleanup)
-
-	ctx.Step(`^the service is running$`, tc.theServiceIsRunning)
-	ctx.Step(`^queue is enabled for payloads$`, tc.queueIsEnabledForJsonnetPayloads)
-	ctx.Step(`^the model endpoint is reachable$`, tc.theModelEndpointIsReachable)
-	ctx.Step(`^there are system providers$`, tc.thereAreSystemProviders)
-	ctx.Step(`^there are system collections$`, tc.thereAreSystemCollections)
-	ctx.Step(`^there is a system collection with id "([^"]*)"$`, tc.thereIsASystemCollectionWithId)
-	ctx.Step(`^the value "([^"]*)" is not empty$`, tc.theValueIsSet)
-	ctx.Step(`^I set the header "([^"]*)" to "([^"]*)"$`, tc.iSetHeaderTo)
-	ctx.Step(`^I unset the header "([^"]*)"$`, tc.iUnsetHeader)
-	ctx.Step(`^I set transaction-id to "([^"]*)"$`, tc.iSetTransactionIdTo)
-	ctx.Step(`^I send a (GET|DELETE|POST|PUT) request to "([^"]*)"$`, tc.iSendARequestTo)
-	ctx.Step(`^I send a (POST|PUT|PATCH) request to "([^"]*)" with body "([^"]*)"$`, tc.iSendARequestToWithBody)
-	ctx.Step(`^I send a (POST|PUT|PATCH) request to "([^"]*)" with body:$`, tc.iSendARequestToWithInlineBody)
-	ctx.Step(`^the response code should be (\d+)$`, tc.theResponseStatusShouldBe)
-	ctx.Step(`^the response code should be (\d+) or (\d+)$`, tc.theResponseStatusShouldBeOr)
-	ctx.Step(`^the response content type should be "([^"]*)"$`, tc.theResponseContentTypeShouldBe)
-	ctx.Step(`^the response body should contain "([^"]*)"$`, tc.theResponseBodyShouldContain)
-	ctx.Step(`^the response should contain "([^"]*)" with value "([^"]*)"$`, tc.theResponseShouldContainWithValue)
-	ctx.Step(`^the response should contain "([^"]*)"$`, tc.theResponseShouldContain)
-	ctx.Step(`^the response should be JSON$`, tc.theResponseShouldBeJSON)
-	ctx.Step(`^the response should contain Prometheus metrics$`, tc.theResponseShouldContainPrometheusMetrics)
-	ctx.Step(`^the metrics should include "([^"]*)"$`, tc.theMetricsShouldInclude)
-	ctx.Step(`^the metrics should show request count for "([^"]*)"$`, tc.theMetricsShouldShowRequestCountFor)
-	// Responses
-	ctx.Step(`^the response should have schema as:$`, tc.theResponseShouldHaveSchemaAs)
-	ctx.Step(`^the response should have schema from file "([^"]*)"$`, tc.theResponseShouldHaveSchemaFromFile)
-	ctx.Step(`^the "([^"]*)" field in the response should be saved as "([^"]*)"$`, tc.theFieldShouldBeSaved)
-	ctx.Step(`^the response should contain the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldContainAtJSONPath)
-	ctx.Step(`^the response should equal the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldEqualAtJSONPath)
-	ctx.Step(`^the response should match the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldMatchAtJSONPath)
-	ctx.Step(`^the response should contain at least the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldContainAtJSONPathAtLeast)
-	ctx.Step(`^the response should not contain the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldNotContainAtJSONPath)
-	ctx.Step(`^the response should not equal the value "([^"]*)" at path "([^"]*)"$`, tc.theResponseShouldNotEqualAtJSONPath)
-	ctx.Step(`^the array at path "([^"]*)" in the response should have length (\d+)$`, tc.theArrayAtPathInResponseShouldHaveLength)
-	ctx.Step(`^the array at path "([^"]*)" in the response should have length "([^"]*)"$`, tc.theArrayAtPathInResponseShouldHaveLength)
-	ctx.Step(`^the array at path "([^"]*)" in the response should have length at least (\d+)$`, tc.theArrayAtPathInResponseShouldHaveLengthAtLeast)
-	ctx.Step(`^the array at path "([^"]*)" in the response should have length at least "([^"]*)"$`, tc.theArrayAtPathInResponseShouldHaveLengthAtLeast)
-	ctx.Step(`^I wait for the evaluation job status to be "([^"]*)"$`, tc.iWaitForEvaluationJobStatus)
-	ctx.Step(`^I set the wait deadline to "([^"]*)"$`, tc.iSetWaitDeadlineTo)
-	ctx.Step(`^I set the wait interval to "([^"]*)"$`, tc.iSetWaitIntervalTo)
-	// Other steps
-	ctx.Step(`^fix this step$`, tc.fixThisStep)
-
-	// MCP-specific steps
-	InitializeMCPSteps(ctx, tc)
-
-	// MLflow artifact steps
-	ctx.Step(`^I fetch the MLflow artifact "([^"]*)" for run "([^"]*)"$`, tc.iFetchMLflowArtifact)
-	ctx.Step(`^I fetch the MLflow artifact "([^"]*)" for experiment "([^"]*)" and job "([^"]*)"$`, tc.iFetchMLflowArtifactByExperimentAndJob)
-	ctx.Step(`^the MLflow artifact should exist$`, tc.theMLflowArtifactShouldExist)
-	ctx.Step(`^the MLflow artifact should contain "([^"]*)"$`, tc.theMLflowArtifactShouldContain)
-	ctx.Step(`^the MLflow artifact should contain the value "([^"]*)" at path "([^"]*)"$`, tc.theMLflowArtifactShouldContainValueAtPath)
-	ctx.Step(`^the MLflow artifact should be valid JSON$`, tc.theMLflowArtifactShouldBeValidJSON)
-	ctx.Step(`^the MLflow artifact field "([^"]*)" should match ISO 8601 format$`, tc.theMLflowArtifactFieldShouldMatchISO8601)
-	ctx.Step(`^I wait for the evaluation job status to match "([^"]*)"$`, tc.iWaitForEvaluationJobStatusToMatch)
-
-	// GPU-specific steps
-	InitializeGPUSteps(ctx, tc)
-
-	// Hardware profile steps (Kubernetes client-go via KUBECONFIG-first FVT helper)
-	InitializeHardwareProfileSteps(ctx, tc)
-}
-
-// --- MLflow Artifact Step Definitions ---
-
-func (tc *scenarioConfig) iFetchMLflowArtifact(artifactPath, runIDPattern string) error {
-	// Resolve run ID from saved values or use literal
-	runID, err := tc.getValue(runIDPattern)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to get run ID from pattern %q: %w", runIDPattern, err))
-	}
-	if runID == "" {
-		return tc.logError(fmt.Errorf("run ID pattern %q resolved to empty string", runIDPattern))
-	}
-
-	// Extract experiment ID from the current response body
-	var experimentID string
-	var respData map[string]interface{}
-	if err := json.Unmarshal(tc.body, &respData); err == nil {
-		if resource, ok := respData["resource"].(map[string]interface{}); ok {
-			if expIDFloat, ok := resource["mlflow_experiment_id"].(float64); ok {
-				experimentID = fmt.Sprintf("%.0f", expIDFloat)
-			} else if expIDStr, ok := resource["mlflow_experiment_id"].(string); ok {
-				experimentID = expIDStr
-			}
-		}
-	}
-	if experimentID == "" {
-		return tc.logError(fmt.Errorf("failed to extract mlflow_experiment_id from response"))
-	}
-
-	return tc.fetchMLflowArtifactWithExperimentID(artifactPath, experimentID, runID)
-}
-
-func (tc *scenarioConfig) fetchMLflowArtifactWithExperimentID(artifactPath, experimentID, runID string) error {
-	baseURL := mlflowBaseURL()
-	workspace := tc.mlflowWorkspace()
-
-	artifactURL := fmt.Sprintf("%s/mlflow/api/2.0/mlflow-artifacts/artifacts/%s/%s/artifacts/%s",
-		baseURL, experimentID, runID, artifactPath)
-
-	tc.logDebug("Fetching MLflow artifact from: %s\n", artifactURL)
-	tc.logDebug("Using workspace: %s\n", workspace)
-
-	client := getMLflowHTTPClient()
-
-	req, err := http.NewRequest("GET", artifactURL, nil)
-	if err != nil {
-		tc.mlflowArtifactError = err
-		return tc.logError(fmt.Errorf("failed to create MLflow artifact request: %w", err))
-	}
-
-	// Add authorization if token is available
-	if authToken := os.Getenv("AUTH_TOKEN"); authToken != "" {
-		req.Header.Set("Authorization", "Bearer "+authToken)
-	}
-
-	// Required for tenant-scoped MLflow access
-	req.Header.Set("X-MLFLOW-WORKSPACE", workspace)
-
-	resp, err := client.Do(req)
-	if err != nil {
-		tc.mlflowArtifactError = err
-		return tc.logError(fmt.Errorf("failed to fetch MLflow artifact: %w", err))
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		tc.mlflowArtifactError = err
-		return tc.logError(fmt.Errorf("failed to read MLflow artifact response: %w", err))
-	}
-
-	if resp.StatusCode != http.StatusOK {
-		tc.mlflowArtifactError = fmt.Errorf("MLflow artifact fetch returned status %d: %s", resp.StatusCode, string(body))
-		return tc.logError(tc.mlflowArtifactError)
-	}
-
-	tc.mlflowArtifactBody = body
-	tc.mlflowArtifactError = nil
-	tc.logDebug("MLflow artifact fetched successfully (%d bytes)\n", len(body))
-	return nil
-}
-
-func (tc *scenarioConfig) theMLflowArtifactShouldExist() error {
-	if tc.mlflowArtifactError != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact fetch failed: %w", tc.mlflowArtifactError))
-	}
-	if len(tc.mlflowArtifactBody) == 0 {
-		return tc.logError(fmt.Errorf("MLflow artifact is empty"))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theMLflowArtifactShouldContain(pathOrString string) error {
-	if tc.mlflowArtifactError != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact fetch failed: %w", tc.mlflowArtifactError))
-	}
-
-	// If it looks like a JSONPath (contains . or [), validate as path
-	if strings.Contains(pathOrString, ".") || strings.Contains(pathOrString, "[") {
-		var jsonData interface{}
-		if err := json.Unmarshal(tc.mlflowArtifactBody, &jsonData); err != nil {
-			return tc.logError(fmt.Errorf("failed to parse MLflow artifact as JSON: %w", err))
-		}
-
-		jsonPath := "$." + pathOrString
-		_, err := jsonpath.Get(jsonPath, jsonData)
-		if err != nil {
-			return tc.logError(fmt.Errorf("MLflow artifact does not contain path %q: %w. Body: %s", pathOrString, err, string(tc.mlflowArtifactBody)))
-		}
-		return nil
-	}
-
-	// Otherwise do literal string search
-	if !strings.Contains(string(tc.mlflowArtifactBody), pathOrString) {
-		return tc.logError(fmt.Errorf("MLflow artifact does not contain %q. Body: %s", pathOrString, string(tc.mlflowArtifactBody)))
-	}
-	return nil
-}
-
-func (tc *scenarioConfig) theMLflowArtifactShouldContainValueAtPath(expected, jsonPath string) error {
-	if tc.mlflowArtifactError != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact fetch failed: %w", tc.mlflowArtifactError))
-	}
-
-	// Parse artifact as JSON
-	var artifactData interface{}
-	if err := json.Unmarshal(tc.mlflowArtifactBody, &artifactData); err != nil {
-		return tc.logError(fmt.Errorf("failed to parse MLflow artifact as JSON: %w", err))
-	}
-
-	// Evaluate JSONPath
-	result, err := jsonpath.Get(jsonPath, artifactData)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to evaluate JSONPath %q on MLflow artifact: %w", jsonPath, err))
-	}
-
-	// Convert result to string for comparison
-	var resultStr string
-	switch v := result.(type) {
-	case string:
-		resultStr = v
-	case float64:
-		resultStr = fmt.Sprintf("%g", v)
-	case bool:
-		resultStr = fmt.Sprintf("%t", v)
-	default:
-		resultBytes, _ := json.Marshal(v)
-		resultStr = string(resultBytes)
-	}
-
-	// Resolve expected value from saved values
-	expectedResolved, err := tc.getValue(expected)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to resolve expected value %q: %w", expected, err))
-	}
-
-	if resultStr != expectedResolved {
-		return tc.logError(fmt.Errorf("MLflow artifact path %q = %q, expected %q", jsonPath, resultStr, expectedResolved))
-	}
-
-	return nil
-}
-
-func (tc *scenarioConfig) iFetchMLflowArtifactByExperimentAndJob(artifactName, experimentID, jobID string) error {
-	// Resolve experiment ID and job ID from saved values
-	experimentIDResolved, err := tc.getValue(experimentID)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to resolve experiment ID %q: %w", experimentID, err))
-	}
-
-	jobIDResolved, err := tc.getValue(jobID)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to resolve job ID %q: %w", jobID, err))
-	}
-
-	// Search for the MLflow run with the matching evaluation_job_id tag
-	baseURL := mlflowBaseURL()
-	workspace := tc.mlflowWorkspace()
-
-	searchURL := fmt.Sprintf("%s/mlflow/api/2.0/mlflow/runs/search", baseURL)
-	searchBody := map[string]interface{}{
-		"experiment_ids": []string{experimentIDResolved},
-		"filter":         fmt.Sprintf("tags.evaluation_job_id = '%s'", jobIDResolved),
-	}
-
-	searchJSON, err := json.Marshal(searchBody)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to marshal search request: %w", err))
-	}
-
-	req, err := http.NewRequest("POST", searchURL, bytes.NewBuffer(searchJSON))
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to create search request: %w", err))
-	}
-
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("X-MLFLOW-WORKSPACE", workspace)
-
-	// Add Authorization header (required for external MLflow access)
-	authToken := os.Getenv("AUTH_TOKEN")
-	if authToken != "" {
-		req.Header.Set("Authorization", "Bearer "+authToken)
-	}
-
-	client := getMLflowHTTPClient()
-	resp, err := client.Do(req)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to search MLflow runs: %w", err))
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to read MLflow search response body: %w", err))
-	}
-	if resp.StatusCode != 200 {
-		return tc.logError(fmt.Errorf("MLflow run search failed with status %d: %s", resp.StatusCode, string(body)))
-	}
-
-	var searchResult map[string]interface{}
-	if err := json.Unmarshal(body, &searchResult); err != nil {
-		return tc.logError(fmt.Errorf("failed to parse search response: %w", err))
-	}
-
-	runs, ok := searchResult["runs"].([]interface{})
-	if !ok || len(runs) == 0 {
-		return tc.logError(fmt.Errorf("no MLflow run found for job %s in experiment %s", jobIDResolved, experimentIDResolved))
-	}
-
-	// Get the run_id from the first match
-	firstRun, ok := runs[0].(map[string]interface{})
-	if !ok {
-		return tc.logError(fmt.Errorf("unexpected MLflow run format: %T", runs[0]))
-	}
-	runInfo, ok := firstRun["info"].(map[string]interface{})
-	if !ok {
-		return tc.logError(fmt.Errorf("MLflow run missing 'info' field or wrong type"))
-	}
-	runID, ok := runInfo["run_id"].(string)
-	if !ok {
-		return tc.logError(fmt.Errorf("MLflow run_id is not a string: %T", runInfo["run_id"]))
-	}
-
-	// Now fetch the artifact using the experiment ID and run_id we already have
-	return tc.fetchMLflowArtifactWithExperimentID(artifactName, experimentIDResolved, runID)
-}
-
-func (tc *scenarioConfig) theMLflowArtifactShouldBeValidJSON() error {
-	if tc.mlflowArtifactError != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact fetch failed: %w", tc.mlflowArtifactError))
-	}
-
-	var jsonData interface{}
-	if err := json.Unmarshal(tc.mlflowArtifactBody, &jsonData); err != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact is not valid JSON: %w. Body: %s", err, string(tc.mlflowArtifactBody)))
-	}
-
-	tc.logDebug("MLflow artifact is valid JSON\n")
-	return nil
-}
-
-func (tc *scenarioConfig) theMLflowArtifactFieldShouldMatchISO8601(fieldPath string) error {
-	if tc.mlflowArtifactError != nil {
-		return tc.logError(fmt.Errorf("MLflow artifact fetch failed: %w", tc.mlflowArtifactError))
-	}
-
-	// Parse artifact as JSON
-	var jsonData interface{}
-	if err := json.Unmarshal(tc.mlflowArtifactBody, &jsonData); err != nil {
-		return tc.logError(fmt.Errorf("failed to parse MLflow artifact as JSON: %w", err))
-	}
-
-	// Extract value using JSONPath
-	jsonPath := "$." + fieldPath
-	results, err := jsonpath.Get(jsonPath, jsonData)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to evaluate JSONPath %q on MLflow artifact: %w", jsonPath, err))
-	}
-
-	resultStr, ok := results.(string)
-	if !ok {
-		return tc.logError(fmt.Errorf("MLflow artifact field %q is not a string: %T", fieldPath, results))
-	}
-
-	// ISO 8601 regex pattern (supports formats like: 2026-07-23T14:30:55.691568184Z, 2026-07-23T14:30:55Z, 2026-07-23T14:30:55+00:00)
-	iso8601Pattern := `^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?$`
-	matched, err := regexp.MatchString(iso8601Pattern, resultStr)
-	if err != nil {
-		return tc.logError(fmt.Errorf("failed to match ISO 8601 pattern: %w", err))
-	}
-
-	if !matched {
-		return tc.logError(fmt.Errorf("MLflow artifact field %q value %q does not match ISO 8601 format", fieldPath, resultStr))
-	}
-
-	tc.logDebug("MLflow artifact field %q is valid ISO 8601: %s\n", fieldPath, resultStr)
-	return nil
-}
-
-func (tc *scenarioConfig) iWaitForEvaluationJobStatusToMatch(statusPattern string) error {
-	// Compile regex once before polling loop for fail-fast validation and performance
-	re, err := regexp.Compile(statusPattern)
-	if err != nil {
-		return tc.logError(fmt.Errorf("invalid status pattern %q: %w", statusPattern, err))
-	}
-
-	deadline := time.Now().Add(tc.waitDeadline)
-	var lastErr error
-	var lastStatus string
-
-	for time.Now().Before(deadline) {
-		// Get current status using {id} pattern which auto-substitutes
-		if err := tc.iSendARequestImpl(http.MethodGet, "/api/v1/evaluations/jobs/{id}", "", "wait for evaluation job status to match pattern"); err != nil {
-			lastErr = err
-			time.Sleep(tc.waitInterval)
-			continue
-		}
-
-		if tc.response != nil && tc.response.StatusCode == http.StatusOK {
-			status, err := tc.getJsonPath("$.status.state")
-			if status != "" {
-				lastStatus = status
-			}
-			if err != nil {
-				lastErr = err
-				time.Sleep(tc.waitInterval)
-				continue
-			}
-
-			// Check if state matches the pattern (supports regex like "completed|failed")
-			if re.MatchString(status) {
-				tc.logDebug("Job reached status matching %q: %s\n", statusPattern, status)
-				return nil
-			}
-
-			tc.logDebug("Waiting for job status to match %q, current: %s\n", statusPattern, status)
-		}
-
-		time.Sleep(tc.waitInterval)
-	}
-
-	if lastErr != nil {
-		return tc.logError(fmt.Errorf("timeout waiting for job status to match %q, last error: %w, last status: %s", statusPattern, lastErr, lastStatus))
-	}
-	return tc.logError(fmt.Errorf("timeout waiting for job status to match %q, last status: %s", statusPattern, lastStatus))
-}
